% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{call_louvain_clusterset}
\alias{call_louvain_clusterset}
\title{Perform Community Detection Using the Louvain Algorithm (internal function)}
\usage{
call_louvain_clusterset(object, resolution = 1)
}
\arguments{
\item{object}{A `ClusterSet` object.}

\item{resolution}{The level of resolution.}
}
\value{
The modified `ClusterSet` object with updated parameters, including 
  the Walktrap algorithm results
}
\description{
Applies the Louvain algorithm for community detection on a graph derived 
from the input data in a `ClusterSet` object. Outputs the resulting clusters 
to a specified file and updates the parameters in the `ClusterSet` object.
}
\details{
This function reads a graph from the input file defined in the `ClusterSet` 
  object, processes the graph using the Louvain community detection algorithm 
  from the `igraph` package, and return the clusters enclose in the `ClusterSet`.
}
\examples{
# Restrict vebosity to info messages only.
set_verbosity(1)
# Load a dataset
load_example_dataset("7871581/files/pbmc3k_medium")
# Select informative genes
res <- select_genes(pbmc3k_medium,
                    distance = "pearson",
                    row_sum=5)

# Cluster informative features
res <- gene_clustering(res, method="closest_neighborhood",
                       algorithm="louvain")

}
