% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{centers}
\alias{centers}
\title{Extract centers from a ClusterSet object.}
\usage{
centers(x)
}
\arguments{
\item{x}{A ClusterSet object.}
}
\description{
Extract centers from a ClusterSet object.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
centers(pbmc3k_medium_clusters)
}
\keyword{internal}
