% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_format_cluster_set}
\alias{check_format_cluster_set}
\title{Check the format of a Clusterset object}
\usage{
check_format_cluster_set(object)
}
\arguments{
\item{object}{the clusterSet object to be tested}
}
\value{
None.
}
\description{
Check the format of a Clusterset object (ncol, nrow, inherits...)
}
\examples{
data("scigenex_test_I1.2")
check_format_cluster_set(scigenex_test_I1.2)
}
\keyword{internal}
