% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{clust_names}
\alias{clust_names}
\title{Names of gene clusters stored in the ClusterSet object}
\usage{
clust_names(x)
}
\arguments{
\item{x}{The ClusterSet object.}
}
\description{
The names of the gene clusters stored in the ClusterSet object.
}
\keyword{internal}
