% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{clust_size,ClusterSet-method}
\alias{clust_size,ClusterSet-method}
\title{Sizes of the clusters stored in a ClusterSet object
clust_size}
\usage{
\S4method{clust_size}{ClusterSet}(x)
}
\description{
The sizes of the clusters stored in a ClusterSet object.
}
