% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{clust_size}
\alias{clust_size}
\title{Sizes of the clusters stored in a ClusterSet object
clust_size}
\usage{
clust_size(x)
}
\arguments{
\item{x}{A ClusterSet object.}
}
\description{
The sizes of the clusters stored in a ClusterSet object.
}
\examples{
# load a dataset
data("scigenex_test_I1.2")
x <- clust_size(scigenex_test_I1.2)
}
\keyword{internal}
