% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{cluster_set_to_xls,ClusterSet-method}
\alias{cluster_set_to_xls,ClusterSet-method}
\title{Write Cluster-Set gene lists into an excel sheet.}
\usage{
\S4method{cluster_set_to_xls}{ClusterSet}(
  object,
  file_path = NULL,
  colname_xls = "official_gene_symbol",
  single_tab = FALSE
)
}
\arguments{
\item{object}{The ClusterSet object.}

\item{file_path}{The file path.}

\item{colname_xls}{The name of the gene column. Default 'official_gene_symbol'.}

\item{single_tab}{Whether all gene lists should be in a single tab. Default to FALSE}
}
\description{
Write gene lists from a Cluster-Set object into an excel sheet.
}
\examples{
#' load_example_dataset('7871581/files/pbmc3k_medium_clusters')
tp_dir <- tempdir()
dir.create(tp_dir, showWarnings = F)
cluster_set_to_xls(pbmc3k_medium_clusters, file.path(tp_dir, "test.xls"))
}
