% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_stats,ClusterSet-method}
\alias{cluster_stats,ClusterSet-method}
\title{Compute statistics about the clusters}
\usage{
\S4method{cluster_stats}{ClusterSet}(object)
}
\arguments{
\item{object}{a ClusterSet object..}
}
\description{
Compute statistics about the clusters
}
\examples{
# load a dataset
data("scigenex_test_I1.2")
df <- cluster_stats(scigenex_test_I1.2)
}
