% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_stats}
\alias{cluster_stats}
\title{Compute statistics about the clusters}
\usage{
cluster_stats(object)
}
\arguments{
\item{object}{a ClusterSet object..}
}
\description{
Compute statistics about the clusters
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
df <- cluster_stats(pbmc3k_medium_clusters)
}
\keyword{internal}
