% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{show,ClusterSet-method}
\alias{show,ClusterSet-method}
\alias{ncol.ClusterSet}
\alias{nrow.ClusterSet}
\alias{clust_names}
\alias{dim,ClusterSet-method}
\alias{col_names,ClusterSet-method}
\alias{row_names,ClusterSet-method}
\alias{[,ClusterSet,ANY,ANY,ANY-method}
\alias{\%in\%,character,ClusterSet-method}
\title{The ClusterSet methods.}
\usage{
\S4method{show}{ClusterSet}(object)

\method{ncol}{ClusterSet}(x)

\method{nrow}{ClusterSet}(x)

clust_names(x)

\S4method{dim}{ClusterSet}(x)

\S4method{col_names}{ClusterSet}(x)

\S4method{row_names}{ClusterSet}(x)

\S4method{[}{ClusterSet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{\%in\%}{character,ClusterSet}(x, table)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{x}{The gene to be searched;}

\item{i}{indices specifying rows to extract. Indices are numeric or character vectors or empty (missing) or NULL.}

\item{j}{indices specifying column to extract. Indices are numeric or character vectors or empty (missing) or NULL.}

\item{...}{See ?'['. Not functionnal here.}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension. Not functionnal here.}

\item{table}{The ClusterSet object.}
}
\description{
The show method of a ClusterSet.

The number of column of a ClusterSet object.

The number of rows of a ClusterSet object.

The names of the gene clusters stored in the ClusterSet object.

The dimension of a ClusterSet object.

The column names of a ClusterSet object.

The row names of a ClusterSet object.

The subsetting operator of a ClusterSet object.

The match operator of a ClusterSet object
}
\section{Functions}{
\itemize{
\item \code{show(ClusterSet)}: The show method of a ClusterSet.

\item \code{ncol(ClusterSet)}: The number of column of a ClusterSet object.

\item \code{nrow(ClusterSet)}: The number of rows of a ClusterSet object.

\item \code{clust_names()}: The names of the gene clusters stored in the ClusterSet object.

\item \code{dim(ClusterSet)}: The dimension of a ClusterSet object.

\item \code{col_names(ClusterSet)}: The column names of a ClusterSet object.

\item \code{row_names(ClusterSet)}: The row names of a ClusterSet object.

\item \code{x[i}: The subsetting operator of a ClusterSet object.
The i axis correspond to clusters and j to column/cells

\item \code{x \%in\% table}: The Match operator of a ClusterSet object

}}
\examples{
   data("scigenex_test_I1.2")               
   x <- ncol(scigenex_test_I1.2)
   x <- nrow(scigenex_test_I1.2)
   x <- dim(scigenex_test_I1.2)
   x <- col_names(scigenex_test_I1.2)
   x <- row_names(scigenex_test_I1.2)
   x <- get_genes(scigenex_test_I1.2)
   x<-  clust_size(scigenex_test_I1.2)
   scigenex_test_I1.2 <- rename(scigenex_test_I1.2)
.   x <- c("gene400", "gene141") \%in\% scigenex_test_I1.2
.   x <- which_clust(scigenex_test_I1.2, genes = c("gene400", "gene141"))
.   scigenex_test_I1.2 <- top_genes(scigenex_test_I1.2, top=5)
.   scigenex_test_I1.2 <- scigenex_test_I1.2[2:3, ]
   clust_names(scigenex_test_I1.2)
   scigenex_test_I1.2 <- rename(scigenex_test_I1.2)
   clust_names(scigenex_test_I1.2)
   scigenex_test_I1.2 <- scigenex_test_I1.2[,col_names(scigenex_test_I1.2)[1:10]]
   show(scigenex_test_I1.2)
}
