% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gene_sets.R
\name{cmp_to_a_list}
\alias{cmp_to_a_list}
\title{Map a user list to a ClusterSet and display various metrics  related to overlap
with internaly stored clusters}
\usage{
cmp_to_a_list(
  object = NULL,
  user_list = NULL,
  name_user_list = "user list",
  background = NULL,
  colors = c("#ED7931", "#3C78AE"),
  module = c("clust_size", "overlap_size", "percent_overlap", "percent_covered",
    "jacard_stat", "hypergeom"),
  as.list = FALSE
)
}
\arguments{
\item{object}{A ClusterSet object}

\item{user_list}{A set of elements to be compared to clusters stored in the object.}

\item{name_user_list}{A name for the user list.}

\item{background}{A background(e.g. all the gene of the genome).}

\item{colors}{A set of colors.}

\item{module}{The analyses to be performed.}

\item{as.list}{Whether a list of plot should be returned (default to a patchwork).}
}
\value{
A patchwork representing the various metrics (or a list of diagram).
}
\description{
Map a user list to a ClusterSet and display various metrics  related to overlap
with internaly stored clusters
}
\examples{
set.seed(124)
load_example_dataset("7871581/files/pbmc3k_medium_clusters")
user_list <- sample(row_names(pbmc3k_medium_clusters), 100)
cmp_to_a_list(pbmc3k_medium_clusters, user_list, background=user_list)
}
