% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{col_names,ClusterSet-method}
\alias{col_names,ClusterSet-method}
\title{Column names of a ClusterSet object.}
\usage{
\S4method{col_names}{ClusterSet}(x)
}
\arguments{
\item{x}{The ClusterSet object}
}
\description{
The column names of a ClusterSet object.
}
