% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{complex9}
\alias{complex9}
\title{The Complex9 dataset
A set of 2D points corresponding to 9 shapes.}
\format{
A data frame with 3031 rows and 3 variables
}
\usage{
data(complex9)
}
\description{
Obtained from Salvador, S. and Chan, P., "Determining the Number of Clusters/Segments 
in Hierarchical clustering/Segmentation Algorithm”, ICTAI 2004,576-584."
Downloaded from:
https://www2.cs.uh.edu/~ml_kdd/restored/Complex&Diamond/complex&diamond.htm
}
\details{
Dataset information:
 ·        Number of Instances: 3031
 ·        Number of Attributes: 2 numeric (X,Y coordinate) and the class
 ·        class: total 9 classes(from 0 to 8)
 ·        Missing Attribute Values: None
 ·        Format:  x, y, (z) class-label

\itemize{
  \item x: x coordinates
  \item y: y coordinates
  \item z: class-label
}
}
\keyword{datasets}
