% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{complex9Noisy}
\alias{complex9Noisy}
\title{The complex9Noisy dataset
A noisy version of the complex9 dataset. A set of 2D points corresponding 
to 9 shapes embedded in a noisy environment. A uniform noise was added to
the dataset.}
\format{
A data frame with 3637 rows and 3 variables
}
\usage{
data(complex9Noisy)
}
\description{
Modified based on Salvador, S. and Chan, P., "Determining the Number of 
Clusters/Segments in Hierarchical clustering/Segmentation Algorithm”, 
ICTAI 2004,576-584." Downloaded from:
https://www2.cs.uh.edu/~ml_kdd/restored/Complex&Diamond/complex&diamond.htm
}
\details{
Dataset information:
 ·        Number of Instances: 3637 (vs 3031 in the original dataset)
 ·        Number of Attributes: 2 numeric (x,y coordinate) and the class
 ·        class: total 10 classes(from 0 to 9). Last one in added noise.
 ·        Missing Attribute Values: None
 ·        Format:  x, y, (z) class-label

\itemize{
  \item x: x coordinates
  \item y: y coordinates
  \item z: class-label
}
}
\keyword{datasets}
