% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{compute_centers,ClusterSet-method}
\alias{compute_centers,ClusterSet-method}
\title{Compute centers of gene modules (e.g mean profiles)}
\usage{
\S4method{compute_centers}{ClusterSet}(object)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}
}
\description{
Compute centers of gene modules (e.g mean profiles)
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
pbmc3k_medium_clusters <- compute_centers(pbmc3k_medium_clusters)
pbmc3k_medium_clusters@dbf_output$center
}
