% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{construct_new_graph}
\alias{construct_new_graph}
\title{Construct a new graph for a ClusterSet object}
\usage{
construct_new_graph(object = obj, k = 5)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{k}{The number of nearest neighbors to consider for each gene.}
}
\value{
A ClusterSet object.
}
\description{
This function constructs a new graph for a given ClusterSet object based on selected genes. 
The graph is constructed using the Density K-Nearest Neighbor (DKNN) method.
}
\examples{
# Set verbosity to 1 to only display info messages.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative features
res <- select_genes(m,
                    distance = "kendall",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Construct a new graph based on genes selected with select_genes()
res <- construct_new_graph(object = res, k = 5)

}
