% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_3_rnd_clust}
\alias{create_3_rnd_clust}
\title{Generate an example dataset with three clusters of profiles}
\usage{
create_3_rnd_clust()
}
\value{
a matrix.
}
\description{
Generate an artificial matrix with with random noise and 3 
clusters of 'expression' profiles (as row).
}
\examples{
m <- create_3_rnd_clust()

}
