% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_4_rnd_clust}
\alias{create_4_rnd_clust}
\title{Generate an example dataset with four clusters of profiles}
\usage{
create_4_rnd_clust()
}
\value{
a matrix.
}
\description{
Generate an artificial matrix with with random noise and 4 
clusters of 'expression' profiles (as row).
}
\examples{
m <- create_4_rnd_clust()


}
