% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gene_clusters.R
\name{DBF}
\alias{DBF}
\title{DBF}
\usage{
DBF(
  data,
  output_path = tempdir(),
  name = NULL,
  dist_threads = 1,
  distance_method = c("pearson", "cosine", "euclidean", "kendall"),
  highest = 0.5,
  k = 100,
  row_sum = 0,
  fdr = 0.05,
  no_dknn_filter = FALSE,
  seed = 123
)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} or \code{Seurat} object.}

\item{output_path}{a character string representing the data directory where
output files will be stored. Default to current working directory.}

\item{name}{a prefix for the names of the intermediary files created by DBF
and MCL.}

\item{dist_threads}{If the "kendall" distance is used, the number of threads (library amap).}

\item{distance_method}{A method to compute the distance to the k-th nearest neighbor 
(one of "pearson", "cosine", "euclidean" or "kendall").}

\item{highest}{During the process, genes will be ordered by their distance to their 
k nearest neighbors (dknn). This parameter controls the fraction of genes with high dknn (ie. noise)
whose neighborhood (i.e associated distances) will be used to compute simulated values. 0 would mean to 
use all the genes.}

\item{k}{the neighborhood size.}

\item{row_sum}{Select only lines whose row sum is greater than row_sum (may be 
-Inf if no filter needed).}

\item{fdr}{a value for the false discovery rate.}

\item{no_dknn_filter}{Do not perform distance to k nearest neighbor (DBF) step.}

\item{seed}{specify seeds for random number generator.}
}
\description{
This function is an internal function used by \code{\link{find_gene_clusters}} to detect
informative elements (\emph{i.e.}, those that belong to dense regions). User
should not use this function. Instead they can use the \code{\link{find_gene_clusters}}
function with \code{clustering} argument set to \code{FALSE}.

See \code{\link{find_gene_clusters}}
}
\section{Warnings}{
 Works only on UNIX-alikes platforms.
}

\references{
Lopez F.,Textoris J., Bergon A., Didier G., Remy E., Granjeaud
S., Imbert J. , Nguyen C. and Puthier D. TranscriptomeBrowser: a powerful
and flexible toolbox to explore productively the transcriptional landscape
of the Gene Expression Omnibus database. PLoSONE, 2008;3(12):e4001.
}
\seealso{
\code{\link{find_gene_clusters}}
}
\author{
Bergon A., Bavais J., Textoris J., Lopez F and Puthier D.
}
\keyword{manip}
