% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{dim,ClusterSet-method}
\alias{dim,ClusterSet-method}
\title{Dimension of a ClusterSet object.
dim}
\usage{
\S4method{dim}{ClusterSet}(x)
}
\arguments{
\item{x}{The ClusterSet object}
}
\description{
The dimension of a ClusterSet object.
}
\keyword{internal}
