% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visium_hull.R
\name{display_hull}
\alias{display_hull}
\title{Draw a hull around a region of interest of Visium data.}
\usage{
display_hull(
  data = NULL,
  ident = NULL,
  size_x = 2.4,
  size_y = 2,
  step_x = 2.6,
  step_y = 2.4,
  delta = 0.3,
  color = "white",
  hull_type = c("walls", "ggforce"),
  expand = unit(0.05, "mm"),
  radius = unit(0.5, "mm"),
  concavity = 1,
  size = 0.7
)
}
\arguments{
\item{data}{A Seurat spatial object (Visium technology).}

\item{ident}{A binary vector whose length correspond to the number
of columns/spots in the seurat object. 1 corresponds to the spots 
of the class of interest for which a hull is to be displayed}

\item{size_x}{The size of the square (x axis)}

\item{size_y}{The size of the square (y axis)}

\item{step_x}{The distance between two points on the x axis.}

\item{step_y}{The distance between two points on the y axis.}

\item{delta}{Add more or less flexibility to search for neighbor points}

\item{color}{A color for the hull.}

\item{hull_type}{The method for drawing the hull: 'ggforce' (see ggforce::geom_mark_hull) or 'wall'. 
If wall, segment wont be connected but this will frequently allows to obtain a better delimited hull.}

\item{expand}{For hull_type 'ggforce'. A numeric or unit vector of length one, specifying the expansion amount.}

\item{radius}{For hull_type 'ggforce'. As expand but specifying the corner radius.}

\item{concavity}{A measure of the concavity of the hull. 1 is very concave while 
it approaches convex as it grows. Defaults to 2.}

\item{size}{The line width.}
}
\value{
A ggplot object (with segments).
}
\description{
Draw a hull around a region of interest of Visium data.
}
\examples{
\dontrun{
library(Seurat)
library(SeuratData)
library(ggplot)
if (! "stxBrain" \%in\% InstalledData()$Dataset) InstallData('stxBrain')
LoadData("stxBrain", type = "anterior1")
# Preprocess the data
anterior1_preprocessed <- preprocess_seurat_data(anterior1, normalization_method = "NormalizeData", approx=F)

# Create a hull around cluster 0
coord_st_data$k <- ifelse(Idents(anterior1_preprocessed)==0, 1, 0)
## Compute the segments of the hull.
path <- visium_hull(coord_st_data, size_x=3.2, size_y=3.6, delta=0.5)
## Add the segments to the ggplot diagram
spatial_plot <- SpatialDimPlot(anterior1_preprocessed, label = TRUE, label.size = 3, pt.size.factor = 1.5)
spatial_plot + display_hull(anterior1_preprocessed)
}
}
\keyword{hull,}
\keyword{spatial}
\keyword{transcriptomics,}
\keyword{visium}
