% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{do_closest_neighbor_graph}
\alias{do_closest_neighbor_graph}
\title{Construct a new graph for a ClusterSet object (internal)}
\usage{
do_closest_neighbor_graph(
  object = NULL,
  k = 5,
  output_path = NULL,
  name = NULL
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{k}{The number of nearest neighbors to consider for each gene.}

\item{output_path}{a character indicating the path where the output files will be stored.}

\item{name}{a character string giving the name for the output files. If NULL, a random name is generated.}
}
\value{
A ClusterSet object.
}
\description{
This function constructs a new graph for a given ClusterSet object based on selected genes. 
The graph is constructed using the Density K-Nearest Neighbor (DKNN) method.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative features
res <- select_genes(m[1:600,],
                    distance = "pearson",
                    k = 75,
                    noise_level = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Construct a new graph based on genes selected with select_genes()
res <- do_closest_neighbor_graph(object = res, k = 5)
}
