% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{do_reciprocal_neighbor_graph}
\alias{do_reciprocal_neighbor_graph}
\title{Construct a graph based on selected genes in a ClusterSet object and their neighbors}
\usage{
do_reciprocal_neighbor_graph(object = NULL, output_path = NULL, name = NULL)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{output_path}{a character indicating the path where the output files will be stored.}

\item{name}{a character string giving the name for the output files. If NULL, a random name is generated.}
}
\value{
The output of this function is a tab-separated file, which serves as input for the MCL algorithm.
}
\description{
This function creates a graph based on the gene selected by select_genes() and their neighbors.
It generates an input file required for the MCL (Markov Cluster) algorithm.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative features
res <- select_genes(m[1:600,],
                    distance = "pearson",
                    k = 75,
                    noise_level = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Construct a graph based on genes selected with select_genes() and their neighbors
do_reciprocal_neighbor_graph(object = res)

}
\keyword{internal}
