% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrich_analysis}
\alias{enrich_analysis}
\alias{enrich_analysis,ClusterSet-method}
\title{Perform gene enrichment analysis.}
\usage{
enrich_analysis(object, specie = "hsapiens")

\S4method{enrich_analysis}{ClusterSet}(object, specie = "hsapiens")
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{specie}{Specie name, as a concatenation of the first letter of the name and the family name, e.g human - hsapien}
}
\value{
A \code{ClusterSet} object
}
\description{
Perform enrichment analysis on all MCL clusters indepentently and store the results in the cluster_annotations slot of the ClusterSet object.
}
\examples{

\dontrun{
## Assuming myobject is a ClusterSet object with at least 1 cluster.

gores <- enrich_analysis(myobject)
}
}
