% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrich_go,ClusterSet-method}
\alias{enrich_go,ClusterSet-method}
\title{Enrichment analysis using GO database of gene clusters from a ClusterSet object}
\usage{
\S4method{enrich_go}{ClusterSet}(object, species = "Hsapiens", ontology = "all")
}
\arguments{
\item{object}{an object of class \code{ClusterSet}.}

\item{species}{a character string indicating the species name,
as a concatenation of the first letter of the name (uppercase) and the family name (lowercase),
 e.g human -> Hsapiens, mouse -> Mmusculus}

\item{ontology}{a character string indicating the ontology to use for the enrichment analysis. One of "BP", "MF", and "CC" ontology, or "ALL".}
}
\value{
A \code{ClusterSet} object
}
\description{
This function performs enrichment analysis of all gene clusters from a ClusterSet object
 and store the results in the cluster_annotations slot.
}
\examples{
# Load a Seurat object
data(pbmc_small, package = "SeuratObject")

# Compute the signatures using find_gene_clusters()
clust_set <- select_genes(pbmc_small,
                          distance = "pearson",
                          no_dknn_filter = TRUE)
                          
# Cluster informative features
clust_set <- gene_clustering(clust_set, 
                            inflation = 2)
# Plot 
clust_set <- top_genes(clust_set)
plot_ggheatmap(clust_set[, names(Idents(pbmc_small))], ident = Idents(pbmc_small))

# Do enrichment analysis using GO ontology
# Only for cluster 1
clust_set <- enrich_go(clust_set[1,])

# Results of enrichment analysis are 'gene_cluster_annotations' slot
print(clust_set@gene_cluster_annotations)

}
