% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrich_viz}
\alias{enrich_viz}
\alias{enrich_viz,ClusterSet-method}
\title{Manhattan-like-plot of enrichment analysis results}
\usage{
enrich_viz(object, clusters = "all", verbose = TRUE)

\S4method{enrich_viz}{ClusterSet}(object, clusters = "all", verbose = TRUE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{clusters}{A vector of cluster id to plot.}

\item{verbose}{Whether or not to print progression in the console.}
}
\value{
A \code{ClusterSet} object
}
\description{
Retrieve enrichment analysis results from a ClusterSet object and draw a Manhattan-like-plot.
}
\examples{

\dontrun{
## Assuming myobject is a ClusterSet object with at least 1 cluster.

enrich_viz(myobject)
}
}
