% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene_clusters.R
\name{filter_by_dot_prod}
\alias{filter_by_dot_prod}
\title{Filter cluster from a ClusterSet object using dot product.}
\usage{
filter_by_dot_prod(object = NULL, av_dot_prod_min = 2)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{av_dot_prod_min}{Any cluster with average dot product below this value is discarded. This allow to delete
clusters in which correlation is influenced/supported by very few samples (typically 1).}
}
\description{
This function filters clusters of gene expression data based 
on their dot products. It aims to remove clusters that have a 
lot of zeros and are supported by only a few cells or spots. 
To do this, the function first converts the gene expression 
data for each cluster into a binary form (values greater than 
1 are set to 1). Then it calculates the dot product for this 
binary matrix, which produces a gene-gene matrix showing the 
number of cells/spots where each pair of genes are expressed 
together. The function then calculates the median value of the 
maximum concordances across all genes, which can be used to 
determine whether a cluster should be filtered out or not.
}
\examples{
# Load a Seurat object
data(pbmc_small, package = "SeuratObject")

# Compute the signatures using find_gene_clusters()
clust_set <- select_genes(pbmc_small,
                          distance = "kendall",
                          k = 75,
                          highest = 0.3,
                          fdr = 1e-8,
                          row_sum = -Inf,
                          no_dknn_filter = TRUE)
                          
# Cluster informative features
clust_set <- gene_clustering(clust_set, 
                            inflation = 1.6,
                            keep_nn = FALSE,
                            k = 5)
                            
# Plot heatmap of gene clusters
plot_heatmap(clust_set)

# Remove the cluster 5 containing less than 4 cells expressing 100\% of the genes in cluster 4
clust_set <- filter_by_dot_prod(clust_set,
                                av_dot_prod_min = 5)
plot_heatmap(clust_set)

}
