% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene_clusters.R
\name{filter_cluster_sd}
\alias{filter_cluster_sd}
\title{Filter out gene clusters with small standard deviation.}
\usage{
filter_cluster_sd(object = NULL, min_sd = 0.1)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{min_sd}{An integer indicating the minimum standard deviation for a clusters to be kept.}
}
\value{
A ClusterSet object where clusters not passing the filter have been removed.
}
\description{
Filter out gene clusters with small standard deviation.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)
}
