% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene_clusters.R
\name{filter_cluster_size}
\alias{filter_cluster_size}
\title{Filter out gene clusters containing few genes}
\usage{
filter_cluster_size(object = NULL, min_cluster_size = 5, rename = FALSE)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{min_cluster_size}{An integer indicating the minimum size for a clusters to be kept.}

\item{rename}{Should cluster be renamed.}
}
\value{
A ClusterSet object where clusters not passing the filter have been removed.
}
\description{
This function filters out gene clusters from a ClusterSet object 
based on their size.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
clust_size(pbmc3k_medium_clusters)
# Remove the cluster with less than 20 genes
clust_set <- filter_cluster_size(pbmc3k_medium_clusters,
                                 min_cluster_size = 20)
clust_size(clust_set)

}
