% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene_clusters.R
\name{filter_gene_clusters}
\alias{filter_gene_clusters}
\title{Filter out gene clusters}
\usage{
filter_gene_clusters(
  object = NULL,
  min_cluster_size = 5,
  min_nb_supporting_cell = 3,
  min_pct_gene_expressed = 50
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{min_cluster_size}{An integer indicating the minimum size of clusters to keep.}

\item{min_nb_supporting_cell}{An integer indicating the minimum number of cell supporting a cluster.
A cell supports a cluster if it expresses at least min_pct_gene_expressed \% of the genes from the cluster.}

\item{min_pct_gene_expressed}{See min_nb_supporting_cell argument.}
}
\value{
A ClusterSet object where clusters that did not pass the filter have been removed.
}
\description{
This function filters out gene clusters from a ClusterSet object
}
