% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cell_clusters.R
\name{find_cell_clusters}
\alias{find_cell_clusters}
\title{Cell clustering}
\usage{
find_cell_clusters(object, min_cluster_size = 5)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{min_cluster_size}{Minimum cluster size.}
}
\value{
A \code{ClusterSet} object.
}
\description{
Identify clusters of cells by cutting the dendrogram generated by the hierarchical clustering algorithm. 
The dendrogram is cut using the hybrid method of the DynamicTreeCut R package which is based on the shape of the dendrogram branches.
For a full description of the algorithm, see Peter Langfelder, Bin Zhang, Steve Horvath, Defining clusters from a hierarchical cluster tree: the Dynamic Tree Cut package for R, Bioinformatics, 2008.
}
\examples{
\dontrun{
## with an artificial dataset
m <- create_4_rnd_clust()

res <- find_gene_clusters(data=m,
                             distance_method="pearson",
                             inflation = 2,
                             k=75,
                             row_sum=-Inf,
                             highest=0.3,
                             min_nb_supporting_cell = 0,
                             fdr = 1e-8)
              
res <- find_cell_clusters(res, min_cluster_size = 3)
res@cell_clust
}
}
