% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gene_clusters.R
\name{find_gene_clusters}
\alias{find_gene_clusters}
\title{The central function of Scigenex package used to select informative genes
and create partitions of co-expressed genes.}
\usage{
find_gene_clusters(
  data = NULL,
  output_path = tempdir(),
  mcl_threads = 1,
  dist_threads = 1,
  name = NULL,
  distance_method = c("pearson", "cosine", "euclidean", "kendall"),
  min_nb_supporting_cell = 2,
  min_pct_gene_expressed = 40,
  min_cluster_size = 10,
  highest = 0.25,
  k = 50,
  row_sum = 0,
  fdr = 0.05,
  inflation = 2,
  no_dknn_filter = FALSE,
  seed = 123
)
}
\arguments{
\item{data}{a \code{matrix}, \code{data.frame} or \code{Seurat} object.}

\item{output_path}{a character string representing the data directory where
output files will be stored. Default to current working directory.}

\item{mcl_threads}{An integer to determine number of threads for mcl algorithm.}

\item{dist_threads}{If the "kendall" distance is used, the number of threads (library amap).}

\item{name}{a prefix for the names of the intermediary files created by DBF
and MCL.}

\item{distance_method}{A method to compute the distance to the k-th nearest neighbor 
(one of "pearson", "cosine", "euclidean" or "kendall").}

\item{min_nb_supporting_cell}{Use to filter out the clusters. The minimum number 
of cell supporting a cluster. A cell supports a cluster if it expresses at least 
min_pct_gene_expressed \% of the genes from the cluster.}

\item{min_pct_gene_expressed}{See min_nb_supporting_cell argument.}

\item{min_cluster_size}{Minimum number of element inside a cluster. 
MCL tend to create lots of clusters with very few (e.g 2) objects.}

\item{highest}{During the process, genes will be ordered by their distance to their 
k nearest neighbors (dknn). This parameter controls the fraction of genes with high dknn (ie. noise)
whose neighborhood (i.e associated distances) will be used to compute simulated values. 0 would mean to 
use all the genes.}

\item{k}{the neighborhood size.}

\item{row_sum}{Select only lines whose row sum is greater than row_sum (may be 
-Inf if no filter needed).}

\item{fdr}{an integer value corresponding to the false discovery rate (range: 0 to 100).}

\item{inflation}{the main control of MCL. Inflation affects cluster
granularity. It is usually chosen somewhere in the range \code{[1.2-3]}.
\code{inflation = 5.0} will tend to result in fine-grained clusterings, and
whereas \code{inflation = 1.2} will tend to result in very coarse grained
clusterings. By default, \code{inflation = 2.0}. Default setting gives very
good results for microarray data when k is set between 70 and 250.}

\item{no_dknn_filter}{Do not perform distance to k nearest neighbor (DBF) step. This may produce
a very large graph.}

\item{seed}{specify seeds for random number generator.}
}
\value{
a ClusterSets class object.
}
\description{
The find_gene_clusters function used an implementation of the  DBFMCL algorithm.
DBFMCL is a tree-steps adaptative algorithm that \emph{(i)} find elements
located in dense areas (DBF), \emph{(ii)} uses selected items to construct a
graph, \emph{(iii)} performs graph partitioning using the Markov CLustering
Algorithm (MCL).

This function requires installation of the mcl program
(\url{http://www.micans.org/mcl}). See "Warnings" section for more
informations.

When analyzing a noisy dataset, one is interested in isolating dense regions
as they are populated with genes/elements that display weak distances to
their nearest neighbors (i.e. strong profile similarities). To isolate these
regions DBF-MCL computes, for each gene/element, the distance with its kth
nearest neighbor (DKNN).In order to define a critical DKNN value that will
depend on the dataset and below which a gene/element will be considered as
falling in a dense area, DBF-MCL computes simulated DKNN values based on 
distances corresponding to genes with the genes highest dknn values. Computed 
distributions of simulated DKNN are used to compute a FDR value for each observed 
DKNN value. Genes with corresponding FDR value <= user defined FDR are selected 
and used to construct a graph. In this graph, edges are constructed between 
two genes (nodes) if one of them belongs to the k-nearest neighbors of the other. 
Edges are weighted based on the respective coefficient of correlation (\emph{i.e.}, 
similarity) and the graph obtained is partitioned using the Markov CLustering 
Algorithm (MCL).
}
\section{Warnings}{
 With the current implementation, this function only works
only on UNIX-like plateforms.

MCL should be installed. One can used the following command lines in a
terminal:

\code{# Download the latest version of mcl (the script has been tested
successfully with the 06-058 version).}
\code{wget http://micans.org/mcl/src/mcl-latest.tar.gz}
\code{# Uncompress and install mcl}
\code{tar xvfz mcl-latest.tar.gz}
\code{cd mcl-xx-xxx}
\code{./configure}
\code{make}
\code{sudo make install}
\code{# You should get mcl in your path}
\code{mcl -h}
}

\examples{

set_verbosity(2)
m <- create_4_rnd_clust()

res <- find_gene_clusters(data=m,
                             distance_method="pearson",
                             inflation = 2,
                             k=75,
                             row_sum=-Inf,
                             highest=0.3,
                             min_nb_supporting_cell = 0,
                             fdr = 1e-8)
plot_heatmap(res, row_labels = F)

}
\references{
- Van Dongen S. (2000) A cluster algorithm for graphs. National
Research Institute for Mathematics and Computer Science in the 1386-3681.
- Lopez F.,Textoris J., Bergon A., Didier G., Remy E., Granjeaud
S., Imbert J. , Nguyen C. and Puthier D. TranscriptomeBrowser: a powerful
and flexible toolbox to explore productively the transcriptional landscape
of the Gene Expression Omnibus database. PLoSONE, 2008;3(12):e4001.
}
\author{
Bergon A., Bavais J., Textoris J., Lopez F and Puthier D.
}
\keyword{MCL.}
\keyword{classification,}
\keyword{clustering,}
\keyword{expression,}
\keyword{gene}
