% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{gene_cluster,ClusterSet-method}
\alias{gene_cluster,ClusterSet-method}
\title{The gene clusters stored in a ClusterSet.
gene_cluster}
\usage{
\S4method{gene_cluster}{ClusterSet}(
  object,
  cluster = NULL,
  as_string = FALSE,
  as_list = FALSE,
  uniqueness = FALSE
)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{cluster}{The cluster of interest. 0 means all cluster. Otherwise a non-null integer value.}

\item{as_string}{Return cluster names as strings.}

\item{as_list}{Return the result as a list.}

\item{uniqueness}{A boolean. If FALSE, any "~[1-9]+$" suffix is removed from gene names to make them unique.}
}
\description{
Returns a named vector (gene as names) and cluster
as value.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
g_clust <- gene_cluster(pbmc3k_medium_clusters)
}
