% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{gene_cluster}
\alias{gene_cluster}
\title{The gene clusters stored in a ClusterSet.
gene_cluster}
\usage{
gene_cluster(object, cluster = 0)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{cluster}{The cluster of interest. 0 means all cluster. Otherwise a non-null integer value.}
}
\description{
Returns a named vector (gene as names) and cluster
as value.
}
\examples{
# load a dataset
data("scigenex_test_I1.2")
g_clust <- gene_cluster(scigenex_test_I1.2)
}
\keyword{internal}
