% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{gene_clustering}
\alias{gene_clustering}
\title{Gene clustering using the Markov Cluster Algorithm (MCL) method.}
\usage{
gene_clustering(
  object = NULL,
  s = 5,
  inflation = 2,
  method = c("closest_neighborhood", "reciprocal_neighborhood"),
  algorithm = c("MCL", "infomap", "louvain", "walktrap"),
  threads = 1,
  output_path = NULL,
  name = NULL,
  keep_nn = FALSE,
  louv_resolution = 5,
  walktrap_step = 4,
  infomap_nb = 10,
  infomap_modularity = TRUE,
  mcl_dir = NULL,
  delete_output_file = FALSE
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{s}{If method="closest_neighborhood", s is an integer value indicating the size of the neighbourhood used for graph construction. Default is 5.}

\item{inflation}{A numeric value indicating the MCL inflation parameter. Default is 2.}

\item{method}{Which method to use to build the graph. If "closest_neighborhood", creates an edge between 
two selected genes a and b if b is part of the kg closest nearest neighbors of a (with kg < k). If "reciprocal_neighborhood"),
inspect the neighborhood of size k of all selected genes and put an edge between two genes a and b if they are reciprocally 
in the neighborhood of the other}

\item{algorithm}{The algorihm to use for clustering: MCL", "infomap", "louvain" or "walktrap".}

\item{threads}{An integer value indicating the number of threads to use for MCL.}

\item{output_path}{a character indicating the path where the output files will be stored.}

\item{name}{a character string giving the name for the output files. If NULL, a random name is generated.}

\item{keep_nn}{Deprecated. Use 'method' instead.}

\item{louv_resolution}{Resolution of Louvain algorithm if chosen.}

\item{walktrap_step}{The length of the random walks to perform for walktrap algorithm.}

\item{infomap_nb}{nb.trials parameter for igraph::cluster_infomap().}

\item{infomap_modularity}{modularity parameter for igraph::cluster_infomap().}

\item{mcl_dir}{A path to MCL (if issue with automated installation).}

\item{delete_output_file}{Whether to delete output files (.graph_input.txt, .graph_out.txt).}
}
\value{
A ClusterSet object
}
\description{
This function performs gene clustering using the MCL algorithm. 
The method starts by creating a graph with genes as nodes and edges connecting each gene to its nearest neighbors.
Then the method use the MCL algorithm to detect clusters of co-expressed genes (method argument).
}
\examples{
# Restrict vebosity to info messages only.
library(Seurat)
set_verbosity(1)

# Load a dataset
load_example_dataset("7871581/files/pbmc3k_medium")

# Select informative genes
res <- select_genes(pbmc3k_medium,
                    distance = "pearson",
                    row_sum=5)
                    
# Cluster informative features

## Method 1 - Construct a graph with a 
## novel neighborhood size
res <- gene_clustering(res, method="closest_neighborhood",
                       inflation = 1.5, threads = 4)
                       
# Display the heatmap of gene clusters
res <- top_genes(res)
plot_heatmap(res)
plot_heatmap(res, cell_clusters = Seurat::Idents(pbmc3k_medium))


## Method 2 - Conserve the same neighborhood 
## size
res <- gene_clustering(res, 
                       inflation = 2.2,
                       method="reciprocal_neighborhood")
                       
# Display the heatmap of gene clusters
res <- top_genes(res)
plot_heatmap(res)
plot_heatmap(res, cell_clusters = Seurat::Idents(pbmc3k_medium))

}
\references{
- Van Dongen S. (2000) A cluster algorithm for graphs. National
Research Institute for Mathematics and Computer Science in the 1386-3681.
}
