% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cells.R
\name{get_cells}
\alias{get_cells}
\title{Extract cell identities from each cell clusters.}
\usage{
get_cells(object, cluster = "all", core = FALSE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{cluster}{A vector of cell cluster identity (all by default).}

\item{core}{A logical to indicate whether to provide core cells.}
}
\value{

}
\description{
Extract cell identities from each cell clusters. Can extract cells from all clusters (default) or from specified clusters.
}
\examples{
\dontrun{
m <- matrix(rnorm(80000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- DBFMCL(data=m,
              distance_method="pearson",
              av_dot_prod_min = 0,
              inflation = 1.2,
              k=25,
              fdr = 10)
res <- cell_clust(res, min_cluster_size = 3)
cells <- get_cells(object = res, cluster="1")
cells
}

}
