% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{get_data_4_DBFMCL}
\alias{get_data_4_DBFMCL}
\title{Title
Fetch an expression matrix from a file, dataframe or Seurat object.}
\usage{
get_data_4_DBFMCL(data = NULL, filename = NULL, path = ".")
}
\arguments{
\item{data}{A \code{matrix}, \code{data.frame} or \code{Seurat} object.}

\item{filename}{A character string representing the file name.}

\item{path}{A character string representing the data directory where
intermediary files are to be stored. Default to current working directory.}
}
\value{
A list containing a matrix and the filename (if filename argument is used).
}
\description{
Title
Fetch an expression matrix from a file, dataframe or Seurat object.
}
