% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_scigenex.R
\name{get_data_for_scigenex}
\alias{get_data_for_scigenex}
\title{Fetch an expression matrix from a file, dataframe or Seurat object.}
\usage{
get_data_for_scigenex(data = NULL, filename = NULL, path = ".")
}
\arguments{
\item{data}{A \code{matrix}, \code{data.frame} or \code{Seurat} object.}

\item{filename}{A character string representing the file name.}

\item{path}{A character string representing the data directory where
intermediary files are to be stored. Default to current working directory.}
}
\value{
A list containing a matrix and the filename (if filename argument is used).
}
\description{
This function fetchs an expression matrix from a file, dataframe or Seurat object.
}
\examples{

\dontrun{
## with an artificial dataset

m <- matrix(rnorm(80000), nc = 20)
res <- get_data_for_scigenex(data=m)
}

}
