% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Extract genes list from each gene clusters.}
\usage{
get_genes(object, cluster = "all", top = FALSE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{cluster}{A vector of gene cluster identity (all by default).}

\item{top}{A logical to indicate whether to provide top genes.}
}
\value{
A character vector.
}
\description{
Extract genes list from each gene clusters. Can extract genes from all clusters (default) or from specified clusters.
}
\examples{
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
# Get all selected genes
x <- get_genes(pbmc3k_medium_clusters)

# Get genes from gene cluster 1
x<- get_genes(pbmc3k_medium_clusters, cluster = 1)

# Get top 5 genes from cluster 1
pbmc3k_medium_clusters <- top_genes(pbmc3k_medium_clusters, top = 5)
x <- get_genes(pbmc3k_medium_clusters, cluster = 1, top = TRUE)
}
