% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_verbosity}
\alias{get_verbosity}
\title{Get the current verbosity level.}
\usage{
get_verbosity()
}
\value{
A vector
}
\description{
This function get the verbosity level of the SciGeneX package which
controls the amount of information that is printed to the console by
the \code{\link{print_msg}} function.
}
\examples{
get_verbosity()

}
