% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_report.R
\name{gm_report}
\alias{gm_report}
\title{Create a report from a ClusterSet and Seurat objects}
\usage{
gm_report(
  cluster_set = NULL,
  seurat_object = NULL,
  seurat_assay = NULL,
  seurat_layer = NULL,
  is_spatial_exp = FALSE,
  report_title = "Gene module report",
  report_subtitle = "An example experiment",
  report_author = "Undefined",
  report_date = format(Sys.time(), "\%d \%B \%Y"),
  out_dir = file.path(fs::path_home(), "gm_book"),
  experimenters = data.frame(),
  workflow_params = data.frame(),
  bioc_org_db = NULL,
  api_key = NULL,
  smp_species = NULL,
  smp_stage = NULL,
  smp_organ = NULL,
  smp_region = NULL,
  rmd_dir = file.path(system.file(package = "scigenex"), "rmarkdown"),
  add_module_score_params = list(nbin = 16, ctrl = 100, name = "MOD_", slot = "data"),
  plot_profiles_params = list(to_lin = TRUE),
  plot_multi_profiles_params = list(legend_name = "Gene\\nModule"),
  FeaturePlot_params = list(cols = RColorBrewer::brewer.pal(3, "BuPu")),
  SpatialFeaturePlot_params = list(pt.size.factor = 1.7),
  SpatialDimPlot_params = list(pt.size.factor = 1.7),
  plot_ggheatmap_params = list(use_top_genes = FALSE, hide_gene_name = TRUE),
  plot_heatmap_params = list(link = "complete", use_top_genes = FALSE, interactive =
    TRUE, line_size_horizontal = 1, label_size = 7),
  cnetplot_params = list(color.params = list(edge = TRUE), showCategory = 6, cex.params =
    list(category_label = 0.6, gene_label = 0.7)),
  rm_tmpdir = TRUE,
  section = c("exp_info", "exp_metadata", "exp_experimenters", "exp_sample",
    "exp_params", "exp_stats", "exp_genes", "exp_heatmap", "exp_dimplot",
    "exp_spatial_dist", "exp_spatial_dimplot", "exp_mean_1", "exp_mean_2",
    "module_spatial", "module_heatmap", "module_iheatmap", "module_umap",
    "module_violin", "module_cell_annot_IA", "module_term_network",
    "module_term_barplot_1", "module_term_barplot_2", "module_term_network_circ",
    "term_table"),
  quiet = FALSE
)
}
\arguments{
\item{cluster_set}{A clusterSet object.}

\item{seurat_object}{Seurat object.}

\item{seurat_assay}{Which assay should be used in the Seurat object.}

\item{seurat_layer}{Which layer should be used in the Seurat object.}

\item{is_spatial_exp}{Whether there are some spatial information in the ClusterSet object.}

\item{report_title}{A title for the report.}

\item{report_subtitle}{A subtitle for the report.}

\item{report_author}{A character string corresponding to one or several authors.}

\item{report_date}{A date.}

\item{out_dir}{A directory where to store the bookdown output.}

\item{experimenters}{A data.frame providing information about experimenters.}

\item{workflow_params}{A data.frame indicating some workflow parameters.}

\item{bioc_org_db}{A gene annotation database as provided by Bioconductor (e.g. org.Hs.eg.db, org.Mm.eg.db, org.Rn.eg.db, o
rg.Dm.eg.db, org.Dr.eg.db, org.Sc.sgd.db, org.Ce.eg.db...). If NULL, module analysis related to functional annotation are canceled.}

\item{api_key}{An API key for Gemini to perform cell type / cell function analysis. If NULL, the 'module_cell_annot_IA' section will be skipped.}

\item{smp_species}{The species of the sample (free text). E.g 'Homo sapiens'.}

\item{smp_stage}{Development stage of the sample (free text). E.g. 'adult'.}

\item{smp_organ}{The sample organ (free text). E.g. 'tonsil'.}

\item{smp_region}{A region in the organ. E.g. 'whole' (which will merge as 'whole tonsil').}

\item{rmd_dir}{A path where to find the templates for creating the book.}

\item{add_module_score_params}{Some parameters for Seurat::add_module_score() function.}

\item{plot_profiles_params}{Some parameters for plot_profiles() function}

\item{plot_multi_profiles_params}{Some parameters for plot_profiles_multi() function.}

\item{FeaturePlot_params}{Some parameters for Seurat::FeaturePlot() function.}

\item{SpatialFeaturePlot_params}{Some parameters for Seurat::SpatialFeaturePlot() function.}

\item{SpatialDimPlot_params}{Some parameters for Seurat::SpatialDimPlot() function.}

\item{plot_ggheatmap_params}{Some parameters for plot_ggheatmap() function.}

\item{plot_heatmap_params}{Some parameters for plot_heatmap() function.}

\item{cnetplot_params}{Some parameters for enrichplot:::cnetplot.enrichResult() function.}

\item{rm_tmpdir}{Whether to delete temporary directory.}

\item{section}{Which section to activate/deactivate.}

\item{quiet}{Whether to run bookdown::render_book() quietly.}
}
\value{
No return value. A report is generated and written to the specified output directory.
}
\description{
Create a report from a ClusterSet experiment and Seurat object. If is_spatial_exp argument is set to TRUE Spatial diagram will be added. 
The function can call Gemini IA to try to guess cell type and functions associated to each gene module.
}
\examples{
library(scigenex)
library(Seurat)
set_verbosity(3)
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
load_example_dataset('7871581/files/pbmc3k_medium')
gm_report(pbmc3k_medium_clusters[1:3,], 
                pbmc3k_medium, 
                smp_species="Homo_sapiens", 
                smp_region="total", 
                smp_organ="blood", 
                bioc_org_db="org.Hs.eg.db",
                api_key=NULL)
set_verbosity(3)
load_example_dataset('7870305/files/lymph_node_tiny_clusters_2')
load_example_dataset('7870305/files/lymph_node_tiny_2')
gm_report(lymph_node_tiny_clusters_2[1:4,], 
                lymph_node_tiny_2, 
                smp_species="Homo sapiens", 
                smp_region="total", 
                smp_organ="lymph node", 
                smp_stage="adult", 
                bioc_org_db="org.Hs.eg.db",
                api_key=NULL,
                is_spatial_exp=TRUE,
                SpatialFeaturePlot_params=list(pt.size.factor = 3000),
                SpatialDimPlot_params=list(pt.size.factor = 3000)) # Object was created with an older seurat version
}
