% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{\%in\%,character,ClusterSet-method}
\alias{\%in\%,character,ClusterSet-method}
\title{Match operator of a ClusterSet object}
\usage{
\S4method{\%in\%}{character,ClusterSet}(x, table)
}
\arguments{
\item{x}{The gene to be searched;}

\item{table}{The ClusterSet object.}
}
\description{
The match operator of a ClusterSet object
}
\keyword{internal}
