% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{grep_clust}
\alias{grep_clust}
\title{Search gene module using a regular expression.}
\usage{
grep_clust(object, reg_exp = NULL)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{reg_exp}{The regular expression indicating the genes to be found.}
}
\description{
Search gene module using a regular expression.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
hit <- grep_clust(pbmc3k_medium_clusters, reg_exp="^T.*[0-9]$")
}
\keyword{internal}
