% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_mcl}
\alias{install_mcl}
\title{Install MCL}
\usage{
install_mcl(force = FALSE)
}
\arguments{
\item{force}{logical indicating whether to force installation even if MCL is
already installed. Default is \code{FALSE}.}
}
\description{
This function installs the MCL (Markov Cluster) program, which is required for
the main functions of the scigenex package. MCL is a cluster
algorithm that uses stochastic flow simulation to cluster graphs.
}
\examples{
# Install MCL
install_mcl()
}
