% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{keep_dbf_graph}
\alias{keep_dbf_graph}
\title{Construct a graph based on selected genes in a ClusterSet object and their neighbors}
\usage{
keep_dbf_graph(object = NULL, output_path = tempdir(), name = NULL)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{output_path}{a character indicating the path where the output files will be stored.}

\item{name}{a character string giving the name for the output files. If NULL, a random name is generated.}
}
\value{
The output of this function is a tab-separated file, which serves as input for the MCL algorithm.
}
\description{
This function creates a graph based on the gene selected by select_genes() and their neighbors.
It generates an input file required for the MCL (Markov Cluster) algorithm.
}
\examples{
# Set verbosity to 1 to only display info messages.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative features
res <- select_genes(m,
                    distance = "kendall",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Construct a graph based on genes selected with select_genes() and their neighbors
keep_dbf_graph(object = res)

}
