% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_example_dataset}
\alias{load_example_dataset}
\title{Load/download a Visium (10X) example dataset.}
\usage{
load_example_dataset(
  dataset = c("lymph_node_tiny", "lymph_node_tiny_clusters"),
  zenodo_id = "7869307"
)
}
\arguments{
\item{dataset}{The name of the dataset.}
}
\value{
Load the dataset.
}
\description{
Load/download a Visium (10X) example dataset.
}
\examples{
# An example Seurat/Visium dataset
load_example_dataset("lymph_node_tiny")
lymph_node_tiny
# An example clusterSet dataset
load_example_dataset("lymph_node_tiny_clusters")
lymph_node_tiny_clusters
}
