% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seurat.R
\name{load_seurat}
\alias{load_seurat}
\alias{load_seurat,ClusterSet-method}
\title{load_seurat}
\usage{
load_seurat(object, seurat_obj = NULL, dimplot_obj = NULL)

\S4method{load_seurat}{ClusterSet}(object, seurat_obj = NULL, dimplot_obj = NULL)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{seurat_obj}{A seurat object to extract cell to group mapping.}

\item{dimplot_obj}{To display cell clusters on the profile diagram with colors extracted from Dimplot output (provide also a Seurat object).}
}
\value{
A ClusterSet object.
}
\description{
Load a seurat object into a ClusterSet object. At the moment the objective is mainly
to store cell identity (i.e cell types/groups to barcode mapping) and
cell type to color mapping.
}
\examples{
# see online examples
}
