% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{lymph_node_selected_spot_2}
\alias{lymph_node_selected_spot_2}
\title{The lymph_node_selected_spot dataset (#2)}
\format{
A vector containing 601 Visium spots.
}
\usage{
data(lymph_node_selected_spot)
}
\description{
A vector with selected spots from 10X Visium Lymph node dataset.
 See https://tinyurl.com/y9hv7vj3
 and https://tinyurl.com/4bwh759j


Dataset information:
 ·        Number of spots: 443
}
\note{
# This dataset was produced using loupe browser (10X genomics)
}
\keyword{datasets}
