% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gene_clusters.R, R/gene_clustering.R
\name{mcl_system_cmd}
\alias{mcl_system_cmd}
\title{Invokes the command line version of Markov CLustering algorithm (MCL).}
\usage{
mcl_system_cmd(object = NULL, inflation = inflation, threads = 1)

mcl_system_cmd(object = NULL, inflation = inflation, threads = 1)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{inflation}{numeric. The inflation parameter used to control the granularity of the clustering.}

\item{threads}{integer. The number of threads to use.}

\item{silent}{if set to TRUE, the progression of the MCL partitionning is
not displayed.}

\item{name}{character vector. The name of the file used to store the graph.}

\item{input_path}{character. The path where the input file is located.}
}
\value{
Returns a file with the ".mcl\_out.txt" extension.

a ClusterSet object with the updated parameters.
}
\description{
This function invokes the mcl system command. MCL is a clustering algorithm
for graphs that was developped by Stijn van Dongen (see references for
further informations).
}
\details{
This function launches the MCL program for graph partitioning. 
MCL is a graph clustering algorithm that detects clusters of nodes in a graph 
based on the flow of information through the edges. 
It is a fast and efficient algorithm that can be used for clustering large-scale graphs.
}
\section{warning}{
 Works only on UNIX-like plateforms. MCL should be
installed. The following command lines can be used for installation.

\code{# Download the latest version of mcl (RTools4TB has been tested
successfully with the 06-058 version).}
\code{wget http://micans.org/mcl/src/mcl-latest.tar.gz}
\code{# Uncompress and install mcl}
\code{tar xvfz mcl-latest.tar.gz}
\code{cd mcl-xx-xxx}
\code{./configure}
\code{make}
\code{sudo make install}
\code{# You should get mcl in your path}
\code{mcl -h}
}

\section{Warnings}{
 With the current implementation, this function only works
only on UNIX-like plateforms.

MCL should be installed. One can used the following command lines in a
terminal:

\code{# Download the latest version of mcl (the script has been tested
successfully with the 06-058 version).}
\code{wget http://micans.org/mcl/src/mcl-latest.tar.gz}
\code{# Uncompress and install mcl}
\code{tar xvfz mcl-latest.tar.gz}
\code{cd mcl-xx-xxx}
\code{./configure}
\code{make}
\code{sudo make install}
\code{# You should get mcl in your path}
\code{mcl -h}
}

\references{
Stijn van Dongen. A cluster algorithm for graphs.  Technical
Report INS-R0010, National Research Institute for Mathematics and Computer
Science in the Netherlands, Amsterdam, May 2000.
\url{http://www.cwi.nl/ftp/CWIreports/INS/INS-R0010.ps.Z}

- Van Dongen S. (2000) A cluster algorithm for graphs. National
Research Institute for Mathematics and Computer Science in the 1386-3681.
}
\author{
Bergon A., Lopez F., Textoris J., Granjeaud S. and Puthier D.
}
\keyword{manip}
