% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_quality_scores.R
\name{module_quality_scores,ClusterSet-method}
\alias{module_quality_scores,ClusterSet-method}
\title{Compute Gene Module Quality Scores}
\usage{
\S4method{module_quality_scores}{ClusterSet}(object)
}
\arguments{
\item{object}{a ClusterSet object.}
}
\value{
A data frame containing the following cluster quality scores:
\itemize{
  \item \code{connectivity}: Connectivity score.
  \item \code{dunn_index}: Dunn index.
  \item \code{silhouette_coefficient}: Mean silhouette coefficient.
}
}
\description{
This function computes various quality metrics (connectivity, dunn index and silhouette coefficient) for gene modules in a ClusterSet object.
}
\examples{
# Load your data object
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
# Compute quality scores of gene modules
quality_scores <- module_quality_scores(pbmc3k_medium_clusters)
quality_scores
}
