% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{ncol.ClusterSet}
\alias{ncol.ClusterSet}
\title{ncol.ClusterSet}
\usage{
\method{ncol}{ClusterSet}(x)
}
\arguments{
\item{x}{The ClusterSet object}
}
\description{
The number of column of a ClusterSet object.
}
\keyword{internal}
