% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefined_dim_ncol_nrow_for_ClusterSet.R
\name{nrow,ClusterSet-method}
\alias{nrow,ClusterSet-method}
\title{nrow}
\usage{
\S4method{nrow}{ClusterSet}(x)
}
\arguments{
\item{x}{A ClusterSet object.}
}
\value{
The number of rows.
}
\description{
The number of rows of a ClusterSet object.
}
\examples{

\dontrun{
  m <- matrix(rnorm(80000), nc=20)
   m[1:100,1:10] <- m[1:100,1:10] + 4
   m[101:200,11:20] <- m[101:200,11:20] + 3
   m[201:300,5:15] <- m[201:300,5:15] + -2
   res <- DBFMCL(data=m,
                 distance_method="pearson",
                 av_dot_prod_min = 0,
                 inflation = 1.2,
                 k=25,
                 fdr = 10)
  nrow(res)
}

}
