% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{nrow.ClusterSet}
\alias{nrow.ClusterSet}
\title{nrow.ClusterSet}
\usage{
nrow.ClusterSet(x)
}
\arguments{
\item{x}{The ClusterSet object}
}
\description{
The number of rows of a ClusterSet object.
}
\keyword{internal}
