% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{plot_clust}
\alias{plot_clust}
\alias{plot_clust,ClusterSet-method}
\title{plot_clust}
\usage{
plot_clust(
  object,
  type = c("line", "tile"),
  to_log2 = FALSE,
  average_only = FALSE,
  average_line_color = "skyblue4",
  standardizing = FALSE,
  ceil = 1,
  floor = -1,
  centering = TRUE
)

\S4method{plot_clust}{ClusterSet}(
  object,
  type = c("line", "tile"),
  to_log2 = FALSE,
  average_only = FALSE,
  average_line_color = "skyblue4",
  standardizing = FALSE,
  ceil = 1,
  floor = -1,
  centering = TRUE
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{type}{The type of diagram ("line or "tile").}

\item{to_log2}{Whether data should be transform in logarithm base 2 (+ 1 as a pseudocount).}

\item{average_only}{Only display the average gene expression profile (if type="line").}

\item{average_line_color}{The color of the average profile.}

\item{standardizing}{Whether rows should be divided by standard deviation.}

\item{ceil}{A value for ceiling (NULL for no ceiling). Ceiling is performed after log transformation, centering and standardization.}

\item{floor}{A floor value (NULL for no ceiling). flooring is performed after log transformation, centering and standardization.}

\item{centering}{Whether rows should be centered.}
}
\value{
A ggplot diagram.
}
\description{
Plot the results (heatmap or profiles) contained in a ClusterSet object.
}
\examples{
# see online examples
}
