% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_clust_enrichments,ClusterSet-method}
\alias{plot_clust_enrichments,ClusterSet-method}
\title{Plot Cluster Enrichments}
\usage{
\S4method{plot_clust_enrichments}{ClusterSet}(
  object,
  stat_shown = c("qvalue", "p.adjust", "pvalue"),
  nb_go_term = 6,
  gradient_palette = colors_for_gradient("Magma"),
  label_fun = NULL,
  term_order = NULL,
  return_mat = FALSE
)
}
\arguments{
\item{object}{An object of class "ClusterSet".}

\item{stat_shown}{A character string specifying the statistic to be shown. Must be one of "qvalue", "p.adjust", or "pvalue". Defaults to "qvalue".}

\item{nb_go_term}{An integer specifying the number of top GO terms to select per cluster. Defaults to 6. Note that some terms may be shared between clusters.}

\item{gradient_palette}{A vector of colors specifying the gradient palette for the color scale. Defaults to colors_for_gradient("Magma").}

\item{label_fun}{A function to customize the labels of the GO terms. Defaults to NULL.}

\item{term_order}{A vector specifying the desired order of the GO terms. Defaults to NULL.}

\item{return_mat}{Logical. Do not return a ggplot object but the prepared matrix.}
}
\value{
A ggplot object displaying the cluster enrichments for the GO terms.
}
\description{
Plot Cluster Enrichments

Display cluster enrichments for gene GO terms.
}
\examples{
load_example_dataset("8028226/files/pbmc3k_medium_clusters_enr_sub")
plot_clust_enrichments(pbmc3k_medium_clusters_enr_sub)
}
\seealso{
\code{\link{enrich_go}}
\code{\link{colors_for_gradient}}
}
