% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{plot_cluster_stats}
\alias{plot_cluster_stats}
\title{Plot the result of cluster_stats()}
\usage{
plot_cluster_stats(x, highlight = NULL)
}
\arguments{
\item{x}{A \code{ClusterSet} object.}

\item{highlight}{a vector with two modalities indicating which clusters to highlight}
}
\value{
A \code{ClusterSet} object.
}
\description{
Plot the result of cluster_stats()
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')

                       
# Compute some statistics about the clusters
df <- cluster_stats(pbmc3k_medium_clusters)  
plot_cluster_stats(df)
plot_cluster_stats(df, highlight=df$size > 8) 

# Select the cluster of interest
pbmc3k_medium_clusters_filtered <- pbmc3k_medium_clusters[df$size > 8, ] 
nclust(pbmc3k_medium_clusters_filtered)  
}
