% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{plot_dist}
\alias{plot_dist}
\title{plot_dist}
\usage{
plot_dist(file_name, path = ".")
}
\arguments{
\item{file_name}{The name of the input file containing distance and threshold values.}

\item{path}{a character string representing the data directory where
input file is stored. Default to current working directory.}
}
\value{
A ggplot diagram.
}
\description{
Plot the observed and simulated distance with the Kth nearest neighbors.
}
\examples{
# see online examples
}
