% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ggheatmap,ClusterSet-method}
\alias{plot_ggheatmap,ClusterSet-method}
\title{Heatmap of data contained in a ClusterSet object.}
\usage{
\S4method{plot_ggheatmap}{ClusterSet}(
  object,
  to_log2 = FALSE,
  use_top_gene = TRUE,
  standardizing = FALSE,
  biotype = "cells",
  ceil = 1,
  floor = -1,
  centering = TRUE
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{to_log2}{Whether data should be transform in logarithm base 2 (+ 0.1 as a pseudocount).}

\item{standardizing}{Whether rows should be divided by standard deviation.}

\item{biotype}{The names of the columns (e.g cells or spots...).}

\item{ceil}{A value for ceiling (NULL for no ceiling). Ceiling is performed after log transformation, centering and standardization.}

\item{floor}{A value for flooring (NULL for no flooring). Flooring is performed after log transformation, centering and standardization.}

\item{centering}{Whether rows should be centered.}

\item{use_top_genes}{A logical to indicate whether to use highly similar genes in the slot top_genes of ClusterSet.}
}
\value{
A ggplot diagram.
}
\description{
Plot the results (heatmap) contained in a ClusterSet object.
}
\examples{
# Todo
}
