% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{plot_markers_to_clusters,ClusterSet-method}
\alias{plot_markers_to_clusters,ClusterSet-method}
\title{Map cell markers to clusters (hypergeometric/jaccard)}
\usage{
\S4method{plot_markers_to_clusters}{ClusterSet}(
  object,
  markers = NULL,
  gradient_palette = c("white", colors_for_gradient("De2")),
  background = NULL,
  cell_type_order = c("None", "hclust"),
  only_specific = FALSE,
  jaccard_scale = 10
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{markers}{A list of cell markers.}

\item{gradient_palette}{A vector of colors for the gradient palette.}

\item{background}{The background for hypergeometric test. Typically the list of genes annotated in the considered gene ontology  (e.g. BP).}

\item{cell_type_order}{The order of cell types. Can be "None" or "hclust". Not supported at the moment.}

\item{only_specific}{Logical value indicating whether to show only specific markers.}

\item{jaccard_scale}{The scale factor for Jaccard coefficient.}
}
\value{
A plot showing the relationship between markers and clusters.
}
\description{
Use jaccard index and hypergeometric to map a list of cell markers to gene clusters.
}
\examples{
load_example_dataset("8028226/files/pbmc3k_medium_clusters_enr")
library(clustermole)
m <- clustermole_markers(species = "hs")
markers <- split(m$gene[m$organ == "Blood" & m$species == "Human"], m$celltype[m$organ == "Blood" & m$species == "Human"])
plot_markers_to_clusters(pbmc3k_medium_clusters_enr, markers=markers)
}
