% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{plot_multi_profiles}
\alias{plot_multi_profiles}
\title{Plot a summarized view of expression profiles.}
\usage{
plot_multi_profiles(
  data = NULL,
  ident = NULL,
  color_cluster = NULL,
  size_text_y = 5,
  size_label = 2,
  legend_name = "Gene\\nModule",
  nb_column = NULL,
  center = FALSE
)
}
\arguments{
\item{data}{A ClusterSet object.}

\item{ident}{A named vector containing the cell type identities for each cell.
Typically the result from the Idents() function on a Seurat object (see Seurat library).}

\item{color_cluster}{A named vector of colors (with cell type as names) used to 
distinguish between different clusters.}

\item{size_text_y}{The font size of the y-axis tick labels.}

\item{size_label}{The font size of the cluster labels.}

\item{legend_name}{A name for the legend.}

\item{nb_column}{The number of columns in the facet grid of the plot. If not provided,
it is automatically computed as the square root of the number of cell types.}

\item{center}{Whether to center (substract mean) each row.}
}
\value{
A ggplot object showing the expression profiles of cell type-specific genes.
}
\description{
This function generates line plots displaying mean expression profiles accross
cell types and modules.
}
\examples{
# Load a Seurat object
load_example_dataset("7871581/files/pbmc3k_medium")
# Load a ClusterSet object
load_example_dataset("7871581/files/pbmc3k_medium_clusters")
                            
plot_multi_profiles(pbmc3k_medium_clusters, ident=Seurat::Idents(pbmc3k_medium))
pal <- discrete_palette(nclust(pbmc3k_medium_clusters))
names(pal) <- names(pbmc3k_medium_clusters@gene_clusters)
plot_multi_profiles(pbmc3k_medium_clusters, 
        ident=Seurat::Idents(pbmc3k_medium), color_cluster = pal)
plot_multi_profiles(pbmc3k_medium_clusters[2:4,], 
      ident=Seurat::Idents(pbmc3k_medium), 
      color_cluster = pal[2:4],
      center=TRUE)

}
