% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial.R
\name{plot_spatial_panel}
\alias{plot_spatial_panel}
\title{A planel of XY scatter plots of Visium data with hexagonal spots.}
\usage{
plot_spatial_panel(
  seurat_obj = NULL,
  genes = NULL,
  metadata = NULL,
  intensity_slot = c("data", "counts", "sct"),
  title = "",
  size_title = 2,
  face_title = c("plain", "italic", "bold", "bold.italic"),
  barwidth = 1,
  barheight = 3,
  axis = TRUE,
  panel_names = NULL,
  ncol_layout = NULL,
  legend = TRUE,
  guides = NULL,
  pt_size = 3.6,
  pt_shape = 6,
  pt_star = TRUE,
  stroke = 0,
  colours = colors_for_gradient("Ju1")
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing spatial expression data.}

\item{genes}{A vector of gene names to plot.}

\item{metadata}{Provide a vector of metadata that will be used instead of genes (i.e. from 
meta.data) slot of a seurat object.}

\item{intensity_slot}{The assay slot to use for the gene expression values.
Must be one of "sct", "counts", or "data". Default is "sct".}

\item{size_title}{The size of the titles.}

\item{face_title}{Font face for the title. Possible values are “plain”, “italic”, “bold” and “bold.italic”.}

\item{barwidth}{A numeric or a grid::unit() object specifying the width of the colourbar. 
Default value is legend.key.width or legend.key.size in theme() or theme.}

\item{barheight}{A numeric or a grid::unit() object specifying the height of the colourbar. 
Default value is legend.key.height or legend.key.size in theme() or theme.}

\item{panel_names}{A vector of panel names to use for each gene plot.}

\item{ncol_layout}{Number of columns to use for the panel layout. Default is the ceiling 
of the number of genes divided by 2.}

\item{legend}{Whether to display a legend for the color scale. Default is FALSE.}

\item{guides}{A string specifying how guides should be treated in the layout. See patchwork::plot_layout().}

\item{pt_size}{The size of the points in the plot. Default is 2.1.}

\item{pt_shape}{The shape of the points in the plot. Default is 16 (a circle).}

\item{pt_star}{A boolean. Whether to use ggstar shapes.}

\item{stroke}{The thickness of margin of points.}

\item{colours}{A vector of colors.}
}
\value{
A ggplot2 object containing the panel of scatter plots.
}
\description{
This function creates a panel of scatter plots for the spatial expression of a list of genes across spots, 
where the X and Y coordinates represent the spatial location of spots and the color represents the 
expression level of the gene.
}
\examples{
library(SeuratData)
library(Seurat)
#InstallData("stxBrain")
set_verbosity(3)
anterior1 <- LoadData("stxBrain", type = "anterior1")
anterior1 <- SCTransform(anterior1, assay = "Spatial")
genes_to_plot <- c("Hpca", "Olig1", "Klf2")
panel <- c("Panel A", "Panel B", "Panel C")
plot_spatial(seurat_obj = anterior1, genes = genes_to_plot, panel_names = panel)
metadata <- c("nCount_Spatial", "nFeature_Spatial", "nCount_SCT", "nFeature_SCT")
plot_spatial_panel(seurat_obj = anterior1, metadata = metadata)
}
