% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_msg}
\alias{print_msg}
\title{Print a message based on the level of verbosity}
\usage{
print_msg(msg, msg_type = "INFO")
}
\arguments{
\item{msg}{The message to be printed}

\item{msg_type}{The type of message, one of "INFO", "DEBUG", or "WARNING"}
}
\value{
None
}
\description{
Print a message based on the level of verbosity
}
\examples{
print_msg("Hello world!", "INFO")
print_msg("Debugging message", "DEBUG")
print_msg("Warning message", "WARNING")
}
