% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_stat}
\alias{print_stat}
\title{Debugging statistics (vector, matrix or dataframe)}
\usage{
print_stat(msg, data, round_val = 2, msg_type = c("INFO", "DEBUG", "WARNING"))
}
\arguments{
\item{msg}{The message to users.}

\item{data}{The vector (numeric) for which the stats are to be produced
(a vector, )}

\item{round_val}{Round the values in its first argument to the specified number
of decimal. Set argument to -1 for no rounding}

\item{msg_type}{The type of message, one of "INFO", "DEBUG", or "WARNING"}
}
\value{
None
}
\description{
Mostly a debugging function that will print some summary
statistics about a numeric vector, matrix or dataframe.
}
\examples{
opt_warn <- options()$warn
print_stat("My data", 1:10, msg_type="INFO")
set_verbosity(3)
print_stat("My data", matrix(rnorm(10), nc=2), msg_type="DEBUG")
set_verbosity(0)
print_stat("My data", matrix(rnorm(10), nc=2), msg_type="DEBUG")
}
\keyword{internal}
