% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{rename_clust}
\alias{rename_clust}
\title{Rename the gene clusters of a ClusterSet}
\usage{
rename_clust(object, new_names = NULL)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{new_names}{The new names for the clusters.}
}
\description{
Rename the gene clusters of a ClusterSet.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
new_obj <- rename_clust(pbmc3k_medium_clusters, new_names = letters[1:nclust(pbmc3k_medium_clusters)])
}
\keyword{internal}
