% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{reorder_clust}
\alias{reorder_clust}
\title{Reorder the clusters from a ClusterSet}
\usage{
reorder_clust(object, new_order = NULL)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{new_order}{The names from the clusterSet in an alternative order.}
}
\description{
Reorder the clusters from a ClusterSet based on their names
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
clust_size(pbmc3k_medium_clusters)
new_obj <- reorder_clust(pbmc3k_medium_clusters, new_order = 15:1)
clust_size(pbmc3k_medium_clusters)
}
\keyword{internal}
