% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_genes.R
\name{reorder_genes}
\alias{reorder_genes}
\title{reorder_genes}
\usage{
reorder_genes(object, order_by = c("gene_names", "hclust", "correlation"))
}
\arguments{
\item{object}{A ClusterSet object.}

\item{order_by}{A character string related to the reordering type. 
Default is "correlation" which reorder genes based on their correlation 
within their respective gene cluster.}
}
\value{
ClusterSet-class object
}
\description{
Reorder genes in gene cluster.
}
\examples{
\dontrun{
m <- create_3_rnd_clust()

res <- find_gene_clusters(data=m,
                             distance_method="pearson",
                             inflation = 2,
                             k=75,
                             row_sum=-Inf,
                             highest=0.3,
                             min_nb_supporting_cell = 0,
                             fdr = 1e-8)
                             
res <- reorder_genes(res, "correlation")

plot_heatmap(res,
             use_core_cells = FALSE,
             use_top_genes = FALSE)

}

}
