% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{row_names,ClusterSet-method}
\alias{row_names,ClusterSet-method}
\title{Row names of a ClusterSet object.
row_names}
\usage{
\S4method{row_names}{ClusterSet}(x)
}
\arguments{
\item{x}{The ClusterSet object}
}
\description{
The row names of a ClusterSet object.
}
