% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{scigenex_test_I1.2}
\alias{scigenex_test_I1.2}
\title{The scigenex_test dataset
A ClusterSet object to be used for unitary tests of scigenex (internal)}
\format{
A ClusterSet object (4 clusters).
}
\usage{
data(scigenex_test_I1.2)
}
\description{
Dataset information:
 ·        Number of clusters: 4
}
\note{
# This dataset was created using the following code:


#Create matrix containing 3 signatures
m <- create_4_rnd_clust()

## Select informative genes
scigenex_test <- select_genes(data=m,
                             distance_method="kendall",
                             k=75,
                             row_sum=-Inf,
                             dist_threads = 6,
                             highest=0.3,
                             fdr = 1e-8)

scigenex_test_I1.2 <- gene_clustering(object = scigenex_test,
                                 inflation = 1.2,
                                 keep_nn = FALSE,
                                 threads = 6,
                                 k = 5)
}
\keyword{datasets}
