% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_genes.R
\name{select_genes}
\alias{select_genes}
\title{Select genes in co-expression patterns}
\usage{
select_genes(
  data = NULL,
  dist_threads = 1,
  distance_method = c("pearson", "cosine", "euclidean", "kendall"),
  highest = 5e-05,
  k = 80,
  row_sum = 1,
  fdr = 0.005,
  which_slot = c("data", "sct", "counts"),
  no_dknn_filter = FALSE,
  seed = 123
)
}
\arguments{
\item{data}{a matrix, data.frame or Seurat object.}

\item{dist_threads}{an integer specifying the number of threads for the kendall correlation.}

\item{distance_method}{a character string indicating the method for computing distances (one of "pearson", "cosine", "euclidean" or "kendall").}

\item{highest}{During the process, genes will be ordered by their distance to their k nearest neighbors (dknn). 
This parameter controls the fraction of genes with high dknn (ie. noise) whose neighborhood (i.e associated distances) 
will be used to compute simulated values. 0 would mean to use all the genes.}

\item{k}{an integer specifying the k nearest neighbours to compute.}

\item{row_sum}{an integer specifying the minimum number of cells expressing a gene. 
Genes expresssed in less than the chosen values are filtered out from the analysis.To keep all genes, use -Inf.}

\item{fdr}{a numeric value indicating the false discovery rate threshold (range: 0 to 100).}

\item{which_slot}{a character string indicating which slot to use from the input scRNA-seq object (one of "data", "sct" or "counts"). 
SCT is the recommended method from Seurat package when working with spatial transcriptomics data.}

\item{no_dknn_filter}{a logical indicating whether to skip the k-nearest-neighbors (KNN) filter. If FALSE, all genes are kept for the next steps.}

\item{seed}{an integer specifying the random seed to use.}
}
\value{
a ClusterSet class object
}
\description{
This function selects genes based on their correlation with other genes. 
It takes as input a normalized gene expression matrix and 
a set of parameters that affect the selection of genes.
}
\examples{

# Set verbosity to 1 to only display info messages.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative genes
res <- select_genes(matrix_test,
                    distance = "kendall",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)

# Display selected genes
res@gene_clusters

}
\references{
- Van Dongen S. (2000) A cluster algorithm for graphs. National
Research Institute for Mathematics and Computer Science in the 1386-3681.
- Lopez F.,Textoris J., Bergon A., Didier G., Remy E., Granjeaud
S., Imbert J. , Nguyen C. and Puthier D. TranscriptomeBrowser: a powerful
and flexible toolbox to explore productively the transcriptional landscape
of the Gene Expression Omnibus database. PLoSONE, 2008;3(12):e4001.
}
\author{
Julie Bavais, Aurelie Bergon, Fabrice Lopez, Julien Textoris, Samuel Granjeaud, Lionel Spinelli and Denis Puthier
}
