% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{show,ClusterSet-method}
\alias{show,ClusterSet-method}
\title{The show method of a ClusterSet.}
\usage{
\S4method{show}{ClusterSet}(object)
}
\arguments{
\item{object}{A ClusterSet object.}
}
\description{
The show method of a ClusterSet.
}
\keyword{internal}
