% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_methods}
\alias{show_methods}
\title{List the method for the ClusterSet object.}
\usage{
show_methods(class = "ClusterSet", where = "scigenex")
}
\arguments{
\item{class}{The name of the class. Default to "ClusterSet".}

\item{where}{The package name. Default to "scigenex".}
}
\value{
A vector with the available methods.
}
\description{
List the method for the ClusterSet object.
}
\examples{
show_methods()
}
