% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2border.R
\name{stratify_seurat}
\alias{stratify_seurat}
\title{Given two spot groups, compute for each spot: (i) the distance of the
closest spot in the other class.}
\usage{
stratify_seurat(
  seurat_obj,
  ident = NULL,
  colours = NULL,
  colours_stratum = NULL,
  pt_size = 4,
  breaks = 5,
  labels = NULL,
  diagnostic_plot = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{ident}{The classes of the spots (0/1). A named vector (name = spot name; value = classes).}

\item{colours}{Colors for the spot classes.}

\item{colours_stratum}{A set of colors for the stratum.}

\item{pt_size}{The size of the points.}

\item{breaks}{The number of breaks.}

\item{labels}{Class labels.}

\item{diagnostic_plot}{Whether to produce a diagnostic diagram. Highly recommanded to visually inspect the results.}
}
\description{
Given two spot groups, compute for each spot: (i) the distance of the
closest spot in the other class.
}
\examples{
library(Seurat)
library(ggplot2)
load_example_dataset("7870305/files/lymph_node_tiny_2")
Seurat::SpatialDimPlot(lymph_node_tiny_2)
identity <- Idents(lymph_node_tiny_2)
classes <- ifelse(Idents(lymph_node_tiny_2) == 7, 1, 0)
names(classes) <- names(identity)
h <- display_hull(lymph_node_tiny_2, ident=classes, color="black", delta=1, size_x=3.4, size_y=3)
plot_spatial(lymph_node_tiny_2, metadata = "seurat_clusters", pt_size = 6) + h
strat <- stratify_seurat(lymph_node_tiny_2, ident=classes)
strat[[1]]
strat[[2]] + ggplot2::coord_equal()
strat[[3]] + ggplot2::coord_equal()
strat[[4]] + ggplot2::coord_equal() 
}
