% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{[,ClusterSet,ANY,ANY,ANY-method}
\alias{[,ClusterSet,ANY,ANY,ANY-method}
\title{Subsetting operator of a ClusterSet object
Extract}
\usage{
\S4method{[}{ClusterSet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{i}{indices specifying rows to extract. Indices are numeric or character vectors or empty (missing) or NULL.}

\item{j}{indices specifying column to extract. Indices are numeric or character vectors or empty (missing) or NULL.}

\item{...}{See ?'['. Not functionnal here.}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension. Not functionnal here.}
}
\description{
The subsetting operator of a ClusterSet object.
The i axis correspond to clusters and j to column/cells
}
\keyword{internal}
