% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{subsample_by_ident,ClusterSet-method}
\alias{subsample_by_ident,ClusterSet-method}
\title{Given ncell, a target number, select ncell from each class of cell/column.}
\usage{
\S4method{subsample_by_ident}{ClusterSet}(object, ident = NULL, nbcell = 1e+06, seed = 123)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{ident}{A named vector. Names are cell/column names, values are classes/identity. 
Typically the result of the Seurat::Ident() function.}

\item{nbcell}{The number of cell to select per population (cell identity).}

\item{seed}{A seed for subsampling.}
}
\description{
Given ncell, a target number, select ncell from each class of cell/column.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
idents <- sample(1:10, size=ncol(pbmc3k_medium_clusters), rep=TRUE)
names(idents) <- col_names(pbmc3k_medium_clusters)
sub <- subsample_by_ident(pbmc3k_medium_clusters, 
         ident=idents,
         nbcell=10)
}
