% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{subsample_by_ident}
\alias{subsample_by_ident}
\title{Given ncell, a target number, select ncell from each class of cell/column.}
\usage{
subsample_by_ident(object, ident = NULL, nbcell = TRUE, seed = 123)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{ident}{A named vector. Names are cell/column names, values are classes/identity. 
Typically the result of the Seurat::Ident() function.}

\item{nbcell}{The number of cell to select.}

\item{seed}{A seed for subsampling.}
}
\description{
Given ncell, a target number, select ncell from each class of cell/column.
}
\examples{
# load a dataset
}
\keyword{internal}
