% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_by_go.R
\name{top_by_go}
\alias{top_by_go}
\title{Select top_genes based on go terms}
\usage{
top_by_go(
  object,
  go_id = "GO:0003677",
  species = "hsapiens",
  gene_id = "hgnc_symbol",
  host = "https://www.ensembl.org"
)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{go_id}{A GO ID.}

\item{species}{The target species according to biomaRt::listDatasets(mart = ensembl)$dataset 
(e.g hsapiens, mmusculus, rnorvegicus...).}

\item{gene_id}{The type of gene identifier in the clusterSet object (default hgnc_symbol). 
use biomaRt::listFilters() to get the vailable list.}

\item{host}{Host to connect to. Defaults to www.ensembl.org.}
}
\value{
A \code{ClusterSet} object.
}
\description{
The clusterSet object contains a top_genes slot that can be used to display 
genes in heatmaps (see \code{plot_heatmap} function). Here the function select 
top_genes based on GO terms.
}
\examples{
\dontrun{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')

# DNA Binding: "GO:0003677"
# Here only for cluster 1 and 2
pbmc3k_medium_clusters <- top_by_go(pbmc3k_medium_clusters[1:2, ], go_id = "GO:0003677")
pbmc3k_medium_clusters@top_genes
}
}
\keyword{internal}
