% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{top_by_grep}
\alias{top_by_grep}
\title{Select top_genes based on a regular expression search}
\usage{
top_by_grep(object, regexp = NULL, as_list = FALSE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{regexp}{A regular expression}

\item{as_list}{Return a list of clusters not a ClusterSet object.}
}
\value{
A \code{ClusterSet} object or a list (see as_list).
}
\description{
The clusterSet object contains a top_genes slot that can be used to display 
genes in heatmaps (see \code{plot_heatmap} function). Here the function select 
top_genes based on a regular expression search
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')

pbmc3k_medium_clusters <- top_by_grep(pbmc3k_medium_clusters, regexp="^CD")
pbmc3k_medium_clusters@top_genes
}
\keyword{internal}
