% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{top_by_intersect}
\alias{top_by_intersect}
\title{Select top_genes based on intersection with a list.}
\usage{
top_by_intersect(object, set = NULL, as_list = FALSE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{set}{A list to compare clusters to.}

\item{as_list}{Return a list of clusters not a ClusterSet object.}
}
\value{
A \code{ClusterSet} object or a list (see as_list).
}
\description{
The clusterSet object contains a top_genes slot that can be used to display 
genes in heatmaps (see \code{plot_heatmap} function). Here the function select 
top_genes based on intersection with a list.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')

set <- c('MS4A1', 'ISG20', 'CD3D', 'SEC14L5', 'RPL11', 'RPL32')
pbmc3k_medium_clusters <- top_by_intersect(pbmc3k_medium_clusters, set=set)
pbmc3k_medium_clusters@top_genes
}
\keyword{internal}
