% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_genes.R
\name{top_genes,ClusterSet-method}
\alias{top_genes,ClusterSet-method}
\title{Most co-expressed genes from each gene cluster}
\usage{
\S4method{top_genes}{ClusterSet}(object, top = 20, cluster = "all")
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{top}{A value for the number of genes to select from each cluster.}

\item{cluster}{A vector of gene cluster identity.}
}
\value{
A \code{ClusterSet} object.
}
\description{
Extract the most highly co-expressed genes of each gene cluster.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')

# Store top genes in the object
pbmc3k_medium_clusters <- top_genes(pbmc3k_medium_clusters)

}
