% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{viz_enrich}
\alias{viz_enrich}
\alias{viz_enrich,ClusterSet-method}
\title{Manhattan-like-plot of enrichment analysis results}
\usage{
viz_enrich(object, clusters = "all", type = "dotplot", verbose = TRUE)

\S4method{viz_enrich}{ClusterSet}(
  object,
  clusters = "all",
  type = c("dotplot", "barplot"),
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{clusters}{A vector of cluster id to plot.}

\item{verbose}{Whether or not to print progression in the console.}
}
\value{
A \code{ClusterSet} object
}
\description{
Retrieve enrichment analysis results from a ClusterSet object and draw a Manhattan-like-plot.
}
\examples{

\dontrun{
## Assuming myobject is a ClusterSet object with at least 1 cluster.

viz_enrich(myobject)
}
}
