% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{which_clust,ClusterSet-method}
\alias{which_clust,ClusterSet-method}
\title{Which clusters contain a set of genes.}
\usage{
\S4method{which_clust}{ClusterSet}(object, genes = NULL)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{genes}{The genes to be searched in the ClusterSet.}
}
\description{
Which clusters contain a set of genes.
}
\examples{
# load a dataset
data("scigenex_test_I1.2")
hit <- which_clust(scigenex_test_I1.2, genes = c("gene400", "gene141"))
}
