% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{write_clust,ClusterSet-method}
\alias{write_clust,ClusterSet-method}
\title{Write the cluster to files.}
\usage{
\S4method{write_clust}{ClusterSet}(
  object,
  sep = ",",
  file_prefix = "scigenex_clust",
  file_suffix = ".csv",
  path = NULL,
  single_file = TRUE,
  write_cname = TRUE
)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{sep}{The separator}

\item{file_prefix}{A file prefix.}

\item{file_suffix}{A file suffix.}

\item{path}{A directory to store the files.}

\item{single_file}{Logical. Whether to write all clusters in a single file (one cluster / line). Need to change the default separator (e.g to ","). The file_prefix is used as file name.}

\item{write_cname}{Whether to add the cluster name. The cluster name is written as a prefix of each line in file(s) and followed by two pipes ("||").}
}
\description{
Write the cluster to files.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
write_clust(pbmc3k_medium_clusters[1:3,], path="/tmp")
}
