% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{write_clust}
\alias{write_clust}
\title{Write the cluster to files.}
\usage{
write_clust(object, sep = "\\n", file_prefix = "scigenex_clust_", path = NULL)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{sep}{The separator}

\item{file_prefix}{A file prefix.}

\item{path}{A directory to store the files.}
}
\description{
Write the cluster to files.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
write_clust(pbmc3k_medium_clusters[1:3,], path="/tmp")
}
\keyword{internal}
