% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{write_dbf}
\alias{write_dbf}
\alias{write_dbf,ClusterSet-method}
\title{write_dbf
Write a ClusterSet into a flat file.}
\usage{
write_dbf(object, filename_out = NULL, path = ".", nb_na_row = 3)

\S4method{write_dbf}{ClusterSet}(object, filename_out = NULL, path = ".", nb_na_row = 5)
}
\arguments{
\item{object}{ClusterSet.}

\item{filename_out}{The outfile name.}

\item{path}{The path to the file.}

\item{nb_na_row}{Number of separating rows (containing NAs).}
}
\value{
Write a file.
}
\description{
write_dbf
Write a ClusterSet into a flat file.
}
