% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_format_cluster_set}
\alias{check_format_cluster_set}
\title{Check the format of a Clusterset object}
\usage{
check_format_cluster_set(object)
}
\arguments{
\item{object}{the clusterSet object to be tested}
}
\value{
None.
}
\description{
Check the format of a Clusterset object (ncol, nrow, inherits...)
}
\examples{
 set_verbosity(0)
 data(pbmc_small, package = "SeuratObject")
 clust_set <- find_gene_clusters(pbmc_small, k=50, no_dknn_filter=TRUE)
 check_format_cluster_set(clust_set)
}
