% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_format_cluster_set}
\alias{check_format_cluster_set}
\title{Check the format of a Clusterset object}
\usage{
check_format_cluster_set(object)
}
\arguments{
\item{object}{the clusterSet object to be tested}
}
\value{
None.
}
\description{
Check the format of a Clusterset object (ncol, nrow, inherits...)
}
\examples{
\dontrun{
 set_verbosity(0)
 data(pbmc_small, package = "SeuratObject")
 clust_set <- select_genes(data=pbmc_small,
                           distance_method="pearson",
                           k=10,
                           row_sum=-Inf,
                           highest=0.95,
                           fdr = 1e-6)
 clust_set <- gene_clustering(object = clust_set,
                              inflation = 1.2,
                              keep_nn = FALSE,
                              k = 5,
                              threads = 1)
 check_format_cluster_set(clust_set)
}
}
