% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{clust_names,ClusterSet-method}
\alias{clust_names,ClusterSet-method}
\title{clust_names}
\usage{
\S4method{clust_names}{ClusterSet}(x)
}
\arguments{
\item{x}{The ClusterSet object}
}
\description{
The names of the gene clusters stored in the ClusterSet object.
}
