% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterset_report.R
\name{cluster_set_report}
\alias{cluster_set_report}
\title{Create a report from a ClusterSet and Seurat object.}
\usage{
cluster_set_report(
  clusterset_object = NULL,
  seurat_object = NULL,
  file_path = NULL,
  force = FALSE,
  report_title = "Scigenex report",
  report_author = "Unknown",
  heatmap_colors = colors_for_gradient("Ju1"),
  heatmap_color_ident = NULL,
  coord_flip = TRUE,
  pt_size = 2.75,
  go_info = c("ID", "Description", "GeneRatio", "pvalue", "qvalue", "Count"),
  verbosity = 1
)
}
\arguments{
\item{clusterset_object}{The ClusterSet object.}

\item{seurat_object}{The Seurat object.}

\item{file_path}{A file path where to store the report (html extension).}

\item{force}{Whether to force erase output file.}

\item{report_title}{A title for the report.}

\item{report_author}{Names of report authors.}

\item{heatmap_colors}{Colors for the heatmap gradient.}

\item{heatmap_color_ident}{Colors for cell classes.}

\item{coord_flip}{Whether to flip spatial coordinates.}

\item{pt_size}{Size of the spots.}

\item{go_info}{Info about GO stats. Should be from the list: "ONTOLOGY", "ID", "Description", 
"GeneRatio", "BgRatio", "pvalue", "p.adjust", "qvalue", "geneID", "Count"}

\item{verbosity}{The verbosity level (0-3, defaut to 1.)}
}
\description{
Create a report from a ClusterSet and Seurat object.
}
\examples{
# Load datasets
# load_example_dataset("8028126/files/pbmc3k_medium_clusters_enr")
# load_example_dataset('7871581/files/pbmc3k_medium')
# cluster_set_report(clusterset_object = pbmc3k_medium_clusters_enr[1:2,],
#                    seurat_object=pbmc3k_medium)
# load_example_dataset("7870305/files/lymph_node_tiny_clusters_2")
# load_example_dataset("7870305/files/lymph_node_tiny_2")
# cluster_set_report(clusterset_object = lymph_node_tiny_clusters_2[1:2,],
                   seurat_object=lymph_node_tiny_2,
                   pt_size=6)
}
