% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{cluster_set_to_xls,ClusterSet-method}
\alias{cluster_set_to_xls,ClusterSet-method}
\title{Write Cluster-Set gene lists into an excel sheet.}
\usage{
\S4method{cluster_set_to_xls}{ClusterSet}(object, file_path = NULL, single_tab = FALSE)
}
\arguments{
\item{object}{The ClusterSet object.}

\item{file_path}{The file path.}

\item{single_tab}{Whether all gene lists should be in a single tab. Default to FALSE}
}
\description{
Write gene lists from a Cluster-Set object into an excel sheet.
}
\examples{
#' Load an example dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')

# Write gene lists to file
tp_dir <- tempdir()
dir.create(tp_dir, showWarnings = FALSE)
cluster_set_to_xls(pbmc3k_medium_clusters, file.path(tp_dir, "test.xls"))
}
