% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{cluster_set_to_xls}
\alias{cluster_set_to_xls}
\title{Write Cluster-Set gene lists into an excel sheet.}
\usage{
cluster_set_to_xls(object, file_path = NULL)
}
\arguments{
\item{object}{The ClusterSet object.}

\item{file_path}{The file path.}
}
\description{
Write gene lists from a Cluster-Set object into an excel sheet.
}
\keyword{internal}
