% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\docType{class}
\name{ClusterSet-class}
\alias{ClusterSet-class}
\title{ClusterSet}
\value{
A ClusterSet object.
}
\description{
This class is a representation of a partitioning algorithm and is intented to store gene clusters.
}
\section{Slots}{

\describe{
\item{\code{data}}{A matrix containing the filtered and partitioned data.}

\item{\code{gene_clusters}}{A list contains the partitioned genes of the dataset. Each element of the list corresponds to a cluster, and contains the indices of the genes assigned to that cluster.}

\item{\code{top_genes}}{A list contains the top genes from the gene clusters. Each element of the list corresponds to a cluster, and contains the indices of the genes assigned to that cluster ranked by their correlation value within their cluster.}

\item{\code{gene_clusters_metadata}}{A list contains metadata related to the gene clusters such as the number of gene clusters, their ID, and the number of genes contained in each of them.}

\item{\code{gene_cluster_annotations}}{A list contains the result obtained from the GO enrichment analysis of gene clusters.}

\item{\code{cell_metadata}}{A list containing metadata related to the cell clusters such as the clustering results the number of cell clusters, their order, colors associated to each cluster,...}

\item{\code{dbf_output}}{A list containing the intermediates outputs of the DBF function : dknn, simulated distances, critical distance and fdr values.}

\item{\code{parameters}}{A list containing the parameter used. Each element of the list correspond to a parameter.}
}}

\examples{

\dontrun{
 m <- matrix(rnorm(80000), nc=20)
 m[1:100,1:10] <- m[1:100,1:10] + 4
 m[101:200,11:20] <- m[101:200,11:20] + 3
 m[201:300,5:15] <- m[201:300,5:15] + -2
 res <- find_gene_clusters(data=m,
                           distance_method="pearson",
                           inflation = 1.2,
                           k=25,
                           fdr = 10)
 is(res)
 res
 }          
}
