% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R, R/ClusterSet_class_save.R
\docType{class}
\name{ClusterSet-class}
\alias{ClusterSet-class}
\title{ClusterSet}
\value{
A ClusterSet object.

A ClusterSet object.
}
\description{
This class is a representation of a partitioning algorithm and is intented to store gene clusters.

This class is a representation of a partitioning algorithm and is intented to store gene clusters.
}
\section{Slots}{

\describe{
\item{\code{data}}{A matrix containing the filtered and partitioned data.}

\item{\code{gene_clusters}}{A list contains the partitioned genes of the dataset. Each element of the list corresponds to a cluster, and contains the indices of the genes assigned to that cluster.}

\item{\code{top_genes}}{A list contains the top genes from the gene clusters. Each element of the list corresponds to a cluster, and contains the indices of the genes assigned to that cluster ranked by their correlation value within their cluster.}

\item{\code{gene_clusters_metadata}}{A list contains metadata related to the gene clusters such as the number of gene clusters, their ID, and the number of genes contained in each of them.}

\item{\code{gene_cluster_annotations}}{A list contains the result obtained from the GO enrichment analysis of gene clusters.}

\item{\code{cell_clusters}}{A list containing the cell clusters.}

\item{\code{cell_metadata}}{A list containing metadata related to the cell clusters such as the number of cell clusters, their order, colors associated to each cluster,...}

\item{\code{dbf_output}}{A list containing the intermediates outputs of the DBF function : dknn, simulated distances, critical distance and fdr values.}

\item{\code{parameters}}{A list containing the parameter used. Each element of the list correspond to a parameter.}

\item{\code{name}}{character. The original input file name (if applicable).}

\item{\code{data}}{matrix. The matrix containing the filtered/partitionned data.}

\item{\code{distances}}{vector. The observed distance values with the knn.}

\item{\code{simulated_distances}}{vector. The simulated distance values with the knn.}

\item{\code{critical_distance}}{vector. The critical distance to k nearest neighbors (dknn) values 
used to select informative genes.}

\item{\code{gene_clusters}}{vector. Mapping of row/genes to gene_clusters.}

\item{\code{cluster_list}}{list. List containing the number of clusters.}

\item{\code{cluster_number}}{vector. Number of gene clusters detected.}

\item{\code{size}}{vector. The size of each cluster.}

\item{\code{fdr}}{The computed FDR for eah gene.}

\item{\code{top_genes}}{matrix The highly co-expressed genes of each gene clusters.}

\item{\code{center}}{matrix. The centers of each clusters.}

\item{\code{parameters}}{list. The parameter used.}

\item{\code{algorithm}}{vector. The algorithm used to produce the clusters.}

\item{\code{cell_clusters}}{list The cell clusters.}

\item{\code{cell_types}}{vector. The cell types.}

\item{\code{cell_colors}}{vector. The cell types to color mapping.}

\item{\code{cell_order}}{vector. How cell should be ordered.}

\item{\code{cluster_annotations}}{list. Functional annotation of clusters.}
}}

\examples{

\dontrun{
  m <- matrix(rnorm(80000), nc=20)
   m[1:100,1:10] <- m[1:100,1:10] + 4
   m[101:200,11:20] <- m[101:200,11:20] + 3
   m[201:300,5:15] <- m[201:300,5:15] + -2
   res <- find_gene_clusters(data=m,
                             distance_method="pearson",
                             av_dot_prod_min = 0,
                             inflation = 1.2,
                             k=25,
                             fdr = 10)
plot_clust(res, ceil = 10, floor = -10)
plot_clust(res, type="tile", ceil = 10, floor = -10)
write_clust(res, filename_out = "ALL.sign.txt")
  is(res)
}
              
}
