% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\docType{class}
\name{ClusterSet-class}
\alias{ClusterSet-class}
\title{ClusterSet-class}
\value{
A ClusterSet object.
}
\description{
This class is a representation of a partitioning algorithm and is intented to store gene clusters.
}
\section{Slots}{

\describe{
\item{\code{data}}{A matrix containing the filtered and partitioned data.}

\item{\code{gene_clusters}}{A list contains the partitioned genes of the dataset. Each element of the list corresponds to a cluster, and contains the indices of the genes assigned to that cluster.}

\item{\code{top_genes}}{A list contains the top genes from the gene clusters. Each element of the list corresponds to a cluster, and contains the indices of the genes assigned to that cluster ranked by their correlation value within their cluster.}

\item{\code{gene_clusters_metadata}}{A list contains metadata related to the gene clusters such as the number of gene clusters, their ID, and the number of genes contained in each of them.}

\item{\code{gene_cluster_annotations}}{A list contains the result obtained from the GO enrichment analysis of gene clusters.}

\item{\code{cells_metadata}}{A list containing metadata related to the cell clusters such as the clustering results the number of cell clusters, their order, colors associated to each cluster,...}

\item{\code{dbf_output}}{A list containing the intermediates outputs of the DBF function : dknn, simulated distances, critical distance and fdr values.}

\item{\code{parameters}}{A list containing the parameter used. Each element of the list correspond to a parameter.}
}}

\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_3_rnd_clust()

# Select informative genes
res <- select_genes(m,
                    distance = "kendall",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Cluster informative features
res <- gene_clustering(res, 
                       inflation = 1.2,
                       keep_nn = FALSE,
                       k = 5)
is(res)

# Plot heatmap of gene clusters
plot_heatmap(res, row_labels = FALSE, line_size_horizontal = 2)
plot_heatmap(res[1,], row_labels = FALSE, line_size_horizontal = 2)
plot_heatmap(res[1:2, ], row_labels = FALSE, line_size_horizontal = 2)
plot_heatmap(res[1:2, 1:15], row_labels = FALSE, line_size_horizontal = 2)

# plot the profiles
idents <- c(rep(1,10), rep(2,10))
names(idents) <- colnames(res@data)
plot_profiles(res,
              ident = idents,
             color_cell_type = c("1" = "#4E79A7", "2" = "#A0CBE8"))
             
# Some methods of the ClusterSet object
x <- ncol(res)
x <- nrow(res)
x <- dim(res)
x <- col_names(res)
x <- row_names(res)
x <- get_genes(res)
x<-  clust_size(res)
x <- c("gene400", "gene141") \%in\% res
x <- which_clust(res, genes = c("gene400", "gene141"))
res <- top_genes(res, top=5)
res <- res[2:3, ]
clust_names(res)
scigenex_test_I1.2 <- rename_clust(res)
clust_names(res)
res <- res[,col_names(res)[1:10]]
show(res)    
}
