% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{show,ClusterSet-method}
\alias{show,ClusterSet-method}
\alias{ncol.ClusterSet}
\alias{nrow.ClusterSet}
\alias{dim,ClusterSet-method}
\alias{col_names,ClusterSet-method}
\alias{row_names,ClusterSet-method}
\alias{[,ClusterSet,ANY,ANY,ANY-method}
\alias{nclust,ClusterSet-method}
\alias{clust_size,ClusterSet-method}
\alias{gene_cluster,ClusterSet-method}
\title{show}
\usage{
\S4method{show}{ClusterSet}(object)

\method{ncol}{ClusterSet}(x)

\method{nrow}{ClusterSet}(x)

\S4method{dim}{ClusterSet}(x)

\S4method{col_names}{ClusterSet}(x)

\S4method{row_names}{ClusterSet}(x)

\S4method{[}{ClusterSet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{nclust}{ClusterSet}(x)

\S4method{clust_size}{ClusterSet}(x)

\S4method{gene_cluster}{ClusterSet}(object, cluster = 0)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{x}{The ClusterSet object}

\item{i}{indices specifying rows to extract. Indices are numeric or character vectors or empty (missing) or NULL.}

\item{j}{indices specifying column to extract. Indices are numeric or character vectors or empty (missing) or NULL.}

\item{...}{See ?'['. Not functionnal here.}

\item{drop}{For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension. Not functionnal here.}

\item{cluster}{The cluster of interest. 0 means all cluster. Otherwise a non-null integer value.}
}
\description{
The show method of a ClusterSet.

The ClusterSet methods.

The number of column of a ClusterSet object.

The number of rows of a ClusterSet object.

The dimension of a ClusterSet object.

The column names of a ClusterSet object.

The row names of a ClusterSet object.

The subsetting operator of a ClusterSet object.

The number of clusters in a ClusterSet object.

The size of the clusters stored in a ClusterSet object.

Returns a named vector (gene as names) and cluster
}
\section{Functions}{
\itemize{
\item \code{show(ClusterSet)}: The number of column of a ClusterSet object.

\item \code{ncol(ClusterSet)}: The number of column of a ClusterSet object.

\item \code{nrow(ClusterSet)}: The number of rows of a ClusterSet object.

\item \code{dim(ClusterSet)}: The dimension of a ClusterSet object.

\item \code{col_names(ClusterSet)}: The column names of a ClusterSet object.

\item \code{row_names(ClusterSet)}: The row names of a ClusterSet object.

\item \code{x[i}: The subsetting operator of a ClusterSet object.
The i axis correspond to clusters and j to column/cells

\item \code{nclust(ClusterSet)}: The number of clusters in a ClusterSet object.

\item \code{clust_size(ClusterSet)}: The size of the clusters stored in a ClusterSet object.

\item \code{gene_cluster(ClusterSet)}: Returns a named vector (gene as names) and cluster
as value.

}}
\examples{
   set_verbosity(0)
   data(pbmc_small, package = "SeuratObject")
   ident <- Seurat::Idents(pbmc_small)
   clust_set <- find_gene_clusters(pbmc_small, k=50, no_dknn_filter=TRUE)
   gene_cluster(clust_set)
   ncol(clust_set)
   nrow(clust_set)
   dim(clust_set)
   row_names(clust_set)
   col_names(clust_set)
   row_names(clust_set)
   clust_names(clust_set)
   clust_size(clust_set)
   show(clust_set)
}
