% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\docType{class}
\name{DBFMCLresult-class}
\alias{DBFMCLresult-class}
\title{Title
This class represents the results of the \code{\link{DBFMCL}} algorithm.}
\value{
A DBFMCLresult object.
}
\description{
Title
This class represents the results of the \code{\link{DBFMCL}} algorithm.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. The original input file name (if applicable).}

\item{\code{data}}{matrix. The matrix containing the filtered/partitionned data.}

\item{\code{cluster}}{vector. Mapping of row/genes to clusters.}

\item{\code{size}}{vector. The size of each cluster.}

\item{\code{center}}{matrix. The centers of each clusters.}

\item{\code{parameters}}{list. The parameter used.}
}}

\examples{

\dontrun{
  m <- matrix(rnorm(80000), nc=20)
  m[1:100,1:10] <- m[1:100,1:10] + 4
  m[101:200,11:20] <- m[101:200,11:20] + 3
  m[201:300,5:15] <- m[201:300,5:15] + -2
  res <- DBFMCL(data=m,
                distance_method="pearson",
                clustering=FALSE,
                k=25)
is(res) 
}
              
}
