% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrich_go}
\alias{enrich_go}
\alias{enrich_go,ClusterSet-method}
\title{Perform gene enrichment analysis.}
\usage{
enrich_go(object, specie = "Hsapiens", ontology = "all", verbose = TRUE)

\S4method{enrich_go}{ClusterSet}(object, specie = "Hsapiens", ontology = "all", verbose = TRUE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{specie}{Specie name, as a concatenation of the first letter of the name (uppercas) and the family name, e.g human - Hsapien}

\item{ontology}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{verbose}{Whether or not to print progression in the console.}
}
\value{
A \code{ClusterSet} object
}
\description{
Perform enrichment analysis on all MCL clusters indepentently and store the results in the cluster_annotations slot of the ClusterSet object.
}
\examples{

\dontrun{
## Assuming myobject is a ClusterSet object with at least 1 cluster.

go_res <- enrich_go(myobject)
}
}
