% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrich_go}
\alias{enrich_go}
\title{Enrichment analysis using GO database of gene clusters from a ClusterSet object}
\usage{
enrich_go(object, species = "Hsapiens", ontology = "all")
}
\arguments{
\item{object}{an object of class \code{ClusterSet}.}

\item{species}{a character string indicating the species name,
as a concatenation of the first letter of the name (uppercase) and the family name (lowercase),
 e.g human -> Hsapiens, mouse -> Mmusculus}

\item{ontology}{a character string indicating the ontology to use for the enrichment analysis. One of "BP", "MF", and "CC" ontology, or "ALL".}
}
\value{
A \code{ClusterSet} object
}
\description{
This function performs enrichment analysis of all gene clusters from a ClusterSet object
 and store the results in the cluster_annotations slot.
}
\keyword{internal}
