% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{enrich_viz}
\alias{enrich_viz}
\alias{enrich_viz,ClusterSet-method}
\title{Manhattan-like-plot of enrichment analysis results}
\usage{
enrich_viz(object, interactive = TRUE, clusters = 1)

\S4method{enrich_viz}{ClusterSet}(object, interactive = TRUE, clusters = 1)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{interactive}{A boolean specifiying if the plot should be interactive.}

\item{clusters}{A vector of cluster names to plot}
}
\value{
A \code{ClusterSet} object
}
\description{
Retrieve enrichment analysis results from a ClusterSet object and draw a Manhattan-like-plot.
}
\examples{

\dontrun{
## Assuming myobject is a ClusterSet object with at least 1 cluster.

enrich_viz(myobject)
}
}
