% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene_clusters.R
\name{filter_cluster_size}
\alias{filter_cluster_size}
\title{Filter out gene clusters containing few genes}
\usage{
filter_cluster_size(object = NULL, min_cluster_size = 5)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{min_cluster_size}{An integer indicating the minimum size for a clusters to be kept.}
}
\value{
A ClusterSet object where clusters not passing the filter have been removed.
}
\description{
This function filters out gene clusters from a ClusterSet object 
based on their size.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative genes
clust_set <- select_genes(m,
                         distance = "pearson",
                         k = 75,
                         highest = 0.3,
                         fdr = 1e-8,
                         row_sum = -Inf)
                    
# Cluster informative features
clust_set <- gene_clustering(clust_set, 
                            inflation = 1.2,
                            keep_nn = FALSE,
                            k = 5)
clust_size(clust_set)

# Remove the cluster with less than 100 genes
clust_set <- filter_cluster_size(clust_set,
                                 min_cluster_size = 100)
clust_size(clust_set)

}
