% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene_clusters.R
\name{filter_cluster_size}
\alias{filter_cluster_size}
\title{Filter out gene clusters based on their size}
\usage{
filter_cluster_size(object = NULL, min_cluster_size = 5)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{min_cluster_size}{An integer indicating the minimum size of clusters to keep.}
}
\value{
A ClusterSet object where clusters that did not pass the filter have been removed.
}
\description{
This function filters out gene clusters from a ClusterSet object based on their size.
}
