% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cell_clusters.R
\name{find_cell_clusters}
\alias{find_cell_clusters}
\title{Cell clustering}
\usage{
find_cell_clusters(object, min_cluster_size = 5)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{min_cluster_size}{Minimum cluster size.}
}
\value{

}
\description{
Identify clusters of cells by cutting the dendrogram generated by the hierarchical clustering algorithm. 
The dendrogram is cut using the hybrid method of the DynamicTreeCut R package which is based on the shape of the dendrogram branches.
For a full description of the algorithm, see Peter Langfelder, Bin Zhang, Steve Horvath, Defining clusters from a hierarchical cluster tree: the Dynamic Tree Cut package for R, Bioinformatics, 2008.
}
\examples{
\dontrun{
## with an artificial dataset
m <- matrix(rnorm(80000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- find_gene_clusters(data=m,
                          distance_method="pearson",
                          av_dot_prod_min = 0,
                          inflation = 1.2,
                          k=25,
                          fdr = 10)
              
res <- find_cell_clusters(res, min_cluster_size = 3)
res@cell_clust
}
}
