% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{gene_cluster,ClusterSet-method}
\alias{gene_cluster,ClusterSet-method}
\title{The gene clusters stored in a ClusterSet.
gene_cluster}
\usage{
\S4method{gene_cluster}{ClusterSet}(object, cluster = 0)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{cluster}{The cluster of interest. 0 means all cluster. Otherwise a non-null integer value.}
}
\description{
Returns a named vector (gene as names) and cluster
as value.
}
