% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cluster.R
\name{gene_cluster}
\alias{gene_cluster}
\title{Extract gene clusters from a cluster set object}
\usage{
gene_cluster(object, cluster = 0, core = FALSE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{cluster}{A vector of gene cluster number (e.g. c(1,3))}
}
\value{
A named character vector.
}
\description{
Extract gene clusters from a cluster set object.
}
\examples{

set_verbosity(0)
data("complex9Noisy")
res <- find_gene_clusters(data=complex9Noisy[ ,1:2],
                          distance_method="euclidean",
                          inflation = 1.25,
                          k=20,
                          min_nb_supporting_cell=0,
                          highest = 0.95,
                          row_sum = -Inf,
                          fdr = 0.0001)

plot(res@data, col=gene_cluster(res))

}
