% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{gene_cluster}
\alias{gene_cluster}
\title{The gene clusters stored in a ClusterSet.
gene_cluster}
\usage{
gene_cluster(object, cluster = 0, as_string = FALSE)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{cluster}{The cluster of interest. 0 means all cluster. Otherwise a non-null integer value.}

\item{as_string}{Return cluster names as strings.}
}
\description{
Returns a named vector (gene as names) and cluster
as value.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
g_clust <- gene_cluster(pbmc3k_medium_clusters)
}
\keyword{internal}
