% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{gene_clustering}
\alias{gene_clustering}
\title{Perform gene clustering using the Markov Cluster Algorithm (MCL) method}
\usage{
gene_clustering(
  object = NULL,
  keep_nn = FALSE,
  k = 5,
  inflation = 2,
  threads = 1
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{keep_nn}{A logical value indicating whether the nearest neighbors graph from select_genes should be kept or not.}

\item{k}{If keep_nn is set to FALSE, k is an integer value indicating the number of nearest neighbors to use in the graph construction. Default is 5.}

\item{inflation}{A numeric value indicating the MCL inflation parameter. Default is 2.}

\item{threads}{An integer value indicating the number of threads to use for MCL.}
}
\value{
A ClusterSet object
}
\description{
This function performs gene clustering using the MCL algorithm. 
The method starts by creating a graph with genes as nodes and edges connecting each gene to its nearest neighbors.
Two methods are available based on keep_nn argument : 
* when using 'keep_nn = FALSE', the method construct a new graph only based on the previously selected genes.
* when using 'keep_nn = TRUE', the method conserve the graph constructed using select_genes with selected genes and their nearest neighbors.
Then the method apply the MCL algorithm to detect clusters of co-expressed genes.
}
\examples{
# Set verbosity to 1 to only display info messages.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative features
res <- select_genes(m,
                    distance = "kendall",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Cluster informative features
res <- gene_clustering(res, 
                       inflation = 1.2,
                       keep_nn = FALSE,
                       k = 5)
# Plot heatmap of feature clusters
plot_heatmap(res)

}
