% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_scigenex.R
\name{get_data_for_scigenex}
\alias{get_data_for_scigenex}
\title{Fetch an expression matrix from a file, dataframe or Seurat object.}
\usage{
get_data_for_scigenex(
  data = NULL,
  layer = c("data", "sct", "counts"),
  assay = NULL
)
}
\arguments{
\item{data}{A \code{matrix}, \code{data.frame} or \code{Seurat} object.}

\item{layer}{One of "SCT", "data" or count. The slot to extract from the seurat object to perform clustering analysis.}

\item{assay}{The assay to use in the Seurat object. If NULL, the function will try to guess.
SCT is the recommended method from Seurat package when working with spatial transcriptomics data.}
}
\value{
An expression matrix.
}
\description{
This function fetchs an expression matrix from a file, dataframe or Seurat object.
}
\examples{
## with an artificial dataset
m <- create_3_rnd_clust()

res <- get_data_for_scigenex(data=m)

}
