% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_scigenex.R
\name{get_data_for_scigenex}
\alias{get_data_for_scigenex}
\title{Fetch an expression matrix from a file, dataframe or Seurat object.}
\usage{
get_data_for_scigenex(data = NULL)
}
\arguments{
\item{data}{A \code{matrix}, \code{data.frame} or \code{Seurat} object.}
}
\value{
An expression matrix.
}
\description{
This function fetchs an expression matrix from a file, dataframe or Seurat object.
}
\examples{

\dontrun{
## with an artificial dataset

m <- create_3_rnd_clust()

res <- get_data_for_scigenex(data=m)
}

}
