% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Extract genes list from each gene clusters.}
\usage{
get_genes(object, cluster = "all", top = FALSE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{cluster}{A vector of gene cluster identity (all by default).}
}
\value{

}
\description{
Extract genes list from each gene clusters. Can extract genes from all clusters (default) or from specified clusters.
}
\examples{

m <- matrix(rnorm(80000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- DBFMCL(data=m,
              distance_method="pearson",
              av_dot_prod_min = 0,
              inflation = 1.2,
              k=25,
              fdr = 10)
              
genes <- get_genes(object = res, cluster="1")
genes

\dontrun{
## with an artificial dataset

m <- matrix(rnorm(80000), nc = 20)
res <- get_data_4_DBFMCL(data=m)
}

}
