% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Extract genes list from each gene clusters.}
\usage{
get_genes(object, cluster = "all", top = FALSE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{cluster}{A vector of gene cluster identity (all by default).}

\item{top}{A logical to indicate whether to provide top genes.}
}
\value{
A character vector.
}
\description{
Extract genes list from each gene clusters. Can extract genes from all clusters (default) or from specified clusters.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative genes
clust_set <- select_genes(m,
                          distance = "kendall",
                          k = 75,
                          highest = 0.3,
                         fdr = 1e-8,
                         row_sum = -Inf)
                    
# Cluster informative features
clust_set <- gene_clustering(clust_set, 
                            inflation = 1.2,
                            keep_nn = FALSE,
                            k = 5)

# Get all selected genes
get_genes(clust_set)

# Get genes from gene cluster 1
get_genes(clust_set, cluster = 1)

# Get top 5 genes from cluster 1
clust_set <- top_genes(clust_set, top = 5)
get_genes(clust_set, cluster = 1, top = TRUE)

}
