% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Extract genes list from each gene clusters.}
\usage{
get_genes(object, cluster = "all", top = FALSE)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{cluster}{A vector of gene cluster identity (all by default).}

\item{top}{A logical to indicate whether to provide top genes.}
}
\value{
A character vector.
}
\description{
Extract genes list from each gene clusters. Can extract genes from all clusters (default) or from specified clusters.
}
\examples{

m <- create_3_rnd_clust()

res <- find_gene_clusters(data=m,
                             distance_method="pearson",
                             inflation = 2,
                             k=75,
                             row_sum=-Inf,
                             highest=0.3,
                             min_nb_supporting_cell = 0,
                             fdr = 1e-8)
              
genes <- get_genes(object = res, cluster="1")
genes

\dontrun{
## with an artificial dataset

m <- matrix(rnorm(80000), nc = 20)
res <- get_data_for_scigenex(data=m)
}

}
