% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visium_hull.R
\name{getFlippedTissueCoordinates}
\alias{getFlippedTissueCoordinates}
\title{Seurat object internally store spot coordinates (see Seurat::GetTissueCoordinates()). 
However, at least in the case of Visium, data are flipped and rotated before SpatialDimPlot. 
This function  return the rotated/flipped tissue Coordinates from a Seurat object.}
\usage{
getFlippedTissueCoordinates(seurat_obj, as_data_frame = FALSE)
}
\arguments{
\item{seurat_obj}{a seurat object with tissue coordinates.}

\item{as_data_frame}{return x/y coords as data.frame. Default to SeuratObject.}
}
\value{
a seurat object with slots/metadata $x_coord and $y_coord (or a dataframe if as_data_frame is TRUE).
}
\description{
Seurat object internally store spot coordinates (see Seurat::GetTissueCoordinates()). 
However, at least in the case of Visium, data are flipped and rotated before SpatialDimPlot. 
This function  return the rotated/flipped tissue Coordinates from a Seurat object.
}
\examples{
load_example_dataset("lymph_node_tiny")
lymph_node_tiny <- getFlippedTissueCoordinates(lymph_node_tiny)
df <- getFlippedTissueCoordinates(lymph_node_tiny, as_data_frame=TRUE)
}
\seealso{
display_visium_hull
}
