% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visium_hull.R
\name{getFlippedTissueCoordinates}
\alias{getFlippedTissueCoordinates}
\title{Seurat object internally store spot coordinates (see Seurat::GetTissueCoordinates()). 
However, at least in the case of Visium, data are flipped and rotated before SpatialDimPlot. 
This function  return the rotated/flipped tissue Coordinates from a Seurat object.}
\usage{
getFlippedTissueCoordinates(seurat_obj, feature = NULL, as_data_frame = FALSE)
}
\arguments{
\item{seurat_obj}{a seurat object with tissue coordinates.}

\item{as_data_frame}{return x/y coords as data.frame. Default to SeuratObject.}
}
\value{
a seurat object with slots $x_coord and $y_coord.
}
\description{
Seurat object internally store spot coordinates (see Seurat::GetTissueCoordinates()). 
However, at least in the case of Visium, data are flipped and rotated before SpatialDimPlot. 
This function  return the rotated/flipped tissue Coordinates from a Seurat object.
}
\examples{
#' ## Install and process the brain dataset
library(SeuratData)
library(Seurat)
#InstallData("stxBrain")
anterior1 <- LoadData("stxBrain", type = "anterior1")
anterior1_df <- getFlippedTissueCoordinates(anterior1, as_data_frame=TRUE)
plot(anterior1_df)
}
\seealso{
display_visium_hull
}
