% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{grep_clust}
\alias{grep_clust}
\title{Search genes within ClusterSet using a REGEXP.}
\usage{
grep_clust(object, regexp = NULL, val = TRUE)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{regexp}{The regular expression}

\item{val}{if FALSE, a vector containing the (integer) indices of the matches determined 
by grep is returned, and if TRUE, a vector containing the matching elements themselves 
is returned.}
}
\description{
Search genes within ClusterSet using a REGEXP.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
grep_clust(pbmc3k_medium_clusters, "[Kk][Rr][Tt]")
}
\keyword{internal}
