% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_example_dataset}
\alias{load_example_dataset}
\title{Load/download a Visium (10X) example dataset.}
\usage{
load_example_dataset(
  dataset = c("7870305/files/lymph_node_tiny_2",
    "7870305/files/lymph_node_tiny_clusters_2", "7869307/files/lymph_node_tiny",
    "7869307/files/lymph_node_tiny_clusters"),
  timeout = NULL
)
}
\arguments{
\item{dataset}{The name of the dataset.}

\item{timeout}{Set the timout for download (options(timeout=timeout))}
}
\value{
Load the dataset.
}
\description{
Load/download a Visium (10X) example dataset.
}
\examples{
# An example Seurat/Visium dataset
load_example_dataset("lymph_node_tiny")
lymph_node_tiny
# An example clusterSet dataset
load_example_dataset("lymph_node_tiny_clusters")
lymph_node_tiny_clusters
}
