% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gene_clusters.R
\name{mcl_system_cmd}
\alias{mcl_system_cmd}
\title{Invokes the command line version of Markov CLustering algorithm (MCL).}
\usage{
mcl_system_cmd(
  name,
  inflation = 2,
  input_path = ".",
  silent = FALSE,
  threads = 1
)
}
\arguments{
\item{name}{a character string corresponding to the file name.}

\item{inflation}{the main control of MCL. Inflation affects cluster
granularity. It is usually chosen somewhere in the range \code{[1.2-5.0]}.
\code{inflation = 5.0} will tend to result in fine-grained clusterings, and
whereas \code{inflation = 1.2} will tend to result in very coarse grained
clusterings. By default, \code{inflation = 2.0}. Default setting gives very
good results for microarray data when k is set around 100.}

\item{input_path}{a character string representing the directory path of 
the input file used by mcl. Default is the current working directory.}

\item{silent}{if set to TRUE, the progression of the MCL partitionning is
not displayed.}

\item{threads}{The number of threads to use.}
}
\value{
Returns a file with the ".mcl\_out.txt" extension.
}
\description{
This function invokes the mcl system command. MCL is a clustering algorithm
for graphs that was developped by Stijn van Dongen (see references for
further informations).
}
\section{warning}{
 Works only on UNIX-like plateforms. MCL should be
installed. The following command lines can be used for installation.

\code{# Download the latest version of mcl (RTools4TB has been tested
successfully with the 06-058 version).}
\code{wget http://micans.org/mcl/src/mcl-latest.tar.gz}
\code{# Uncompress and install mcl}
\code{tar xvfz mcl-latest.tar.gz}
\code{cd mcl-xx-xxx}
\code{./configure}
\code{make}
\code{sudo make install}
\code{# You should get mcl in your path}
\code{mcl -h}
}

\references{
Stijn van Dongen. A cluster algorithm for graphs.  Technical
Report INS-R0010, National Research Institute for Mathematics and Computer
Science in the Netherlands, Amsterdam, May 2000.
\url{http://www.cwi.nl/ftp/CWIreports/INS/INS-R0010.ps.Z}
}
\author{
Bergon A., Lopez F., Textoris J., Granjeaud S. and Puthier D.
}
\keyword{manip}
