% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{mcl_system_cmd}
\alias{mcl_system_cmd}
\title{Run the MCL program for graph partitioning.}
\usage{
mcl_system_cmd(object = NULL, inflation = inflation, threads = 1)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{inflation}{numeric. The inflation parameter used to control the granularity of the clustering.}

\item{threads}{integer. The number of threads to use.}

\item{name}{character vector. The name of the file used to store the graph.}

\item{input_path}{character. The path where the input file is located.}
}
\value{
a ClusterSet object with the updated parameters.
}
\description{
Run the MCL program for graph partitioning.
}
\details{
This function launches the MCL program for graph partitioning. 
MCL is a graph clustering algorithm that detects clusters of nodes in a graph 
based on the flow of information through the edges. 
It is a fast and efficient algorithm that can be used for clustering large-scale graphs.
}
\section{Warnings}{
 With the current implementation, this function only works
only on UNIX-like plateforms.

MCL should be installed. One can used the following command lines in a
terminal:

\code{# Download the latest version of mcl (the script has been tested
successfully with the 06-058 version).}
\code{wget http://micans.org/mcl/src/mcl-latest.tar.gz}
\code{# Uncompress and install mcl}
\code{tar xvfz mcl-latest.tar.gz}
\code{cd mcl-xx-xxx}
\code{./configure}
\code{make}
\code{sudo make install}
\code{# You should get mcl in your path}
\code{mcl -h}
}

\references{
- Van Dongen S. (2000) A cluster algorithm for graphs. National
Research Institute for Mathematics and Computer Science in the 1386-3681.
}
