% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{mcl_system_cmd}
\alias{mcl_system_cmd}
\title{Call MCL program for graph partitioning (internal).}
\usage{
mcl_system_cmd(object = NULL, inflation = inflation, threads = 1)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{inflation}{numeric. The inflation parameter used to control the granularity of the clustering.}

\item{threads}{integer. The number of threads to use.}
}
\value{
a ClusterSet object with the updated parameters.
}
\description{
Call MCL program for graph partitioning (internal).
}
\details{
This function call the MCL program for graph partitioning. 
MCL is a graph clustering algorithm that detects clusters of nodes in a graph 
based on the flow of information through the edges. 
It is a fast and efficient algorithm that can be used for clustering large-scale graphs.
}
\section{Warnings}{
 With the current implementation, this function only works
only on UNIX-like plateforms.

MCL should be installed. Run the \code{install_mcl} command to install it.
}

\references{
- Van Dongen S. (2000) A cluster algorithm for graphs. National
Research Institute for Mathematics and Computer Science in the 1386-3681.
}
