% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{nclust}
\alias{nclust}
\title{Number of clusters in a ClusterSet object.
nclust}
\usage{
nclust(x)
}
\arguments{
\item{x}{The ClusterSet object}
}
\description{
The number of clusters in a ClusterSet object.
}
\examples{
# load a dataset
data("scigenex_test_I1.2")
n_clust <- nclust(scigenex_test_I1.2)
}
\keyword{internal}
