% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{plot_cluster_stats}
\alias{plot_cluster_stats}
\title{Plot the result of cluster_stats()}
\usage{
plot_cluster_stats(x, highlight = NULL)
}
\arguments{
\item{x}{A \code{ClusterSet} object.}

\item{highlight}{a vector with two modalities indicating which clusters to highlight}
}
\value{
A \code{ClusterSet} object.
}
\description{
Plot the result of cluster_stats()
}
\examples{
# Set verbosity to 1 to only display info messages.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative genes
res <- select_genes(m,
                    distance = "pearson",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Cluster informative features

res <- gene_clustering(res, 
                       inflation = 1.2,
                       keep_nn = FALSE,
                       k = 5)
                       
# Compute some statistics about the clusters
df <- cluster_stats(res)  
plot_cluster_stats(df)
plot_cluster_stats(df, highlight=df$sd_total > 1.5) 

# Select the cluster of interest
res <- res[df$sd_total > 1.5, ] 
nclust(res)          
}
