% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{plot_dbf}
\alias{plot_dbf}
\alias{plot_dbf,DBFMCLresult-method}
\title{Title
Plot the results (heatmap or profiles) contained in a DBFMCLresult object.}
\usage{
plot_dbf(
  object,
  type = c("line", "tile"),
  to_log2 = TRUE,
  color_palette = NULL,
  standardizing = FALSE,
  centering = TRUE
)

\S4method{plot_dbf}{DBFMCLresult}(
  object,
  type = c("line", "tile"),
  to_log2 = TRUE,
 
    color_palette = "#0000BF,#0000FF,#0080FF,#00FFFF,#40FFBF,#80FF80,#BFFF40,#FFFF00,#FF8000,#FF0000,#BF0000",
  standardizing = FALSE,
  centering = TRUE
)
}
\arguments{
\item{object}{A DBFMCLresult object.}

\item{type}{The type of diagram.}

\item{to_log2}{Whether data should be transform in logarithm base 2 (+ 1 as a pseudocount).}

\item{color_palette}{A color palette.}

\item{standardizing}{Whether rows should be divided by standard deviation.}

\item{centering}{Whether rows should be centered.}
}
\value{
A ggplot diagram.
}
\description{
Title
Plot the results (heatmap or profiles) contained in a DBFMCLresult object.
}
