% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dist}
\alias{plot_dist}
\title{plot_dist}
\usage{
plot_dist(object, path = ".")
}
\arguments{
\item{object}{A ClusterSet object.}

\item{path}{A character string representing the data directory where
input file containing distances and cutting threshold is stored. 
Default to current working directory.}
}
\value{
A ggplot diagram.
}
\description{
Plot the observed and simulated distance with the Kth nearest neighbors.
}
\examples{
# see online examples
}
