% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot Distribution of distances}
\usage{
plot_dist(
  object,
  bins = 200,
  alpha = 0.5,
  colors = c(Simulated = "blue", Observed = "red"),
  xlim = NULL
)
}
\arguments{
\item{object}{A "ClusterSet"object.}

\item{bins}{The number of bins to use in the histogram (default is 200).}

\item{alpha}{Transparency of the histogram bars (default is 0.5).}

\item{colors}{A named vector of colors to use for the histograms, with "Simulated" and "Observed" as the named elements (default is c("Simulated"="blue", "Observed"="red")).}

\item{xlim}{Limits for the x axis of the histogram (e.g. c(0.8, 1)).}
}
\value{
A ggplot object.
}
\description{
This function creates a histogram of the simulated and observed distances of a Seurat clusterSet object.
}
