% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot Distribution of distances}
\usage{
plot_dist(
  object,
  bins = 150,
  alpha = 0.5,
  colors = c(Simulated = "#FB8500", Observed = "#36949D"),
  xlim = NULL,
  vline_color = "#4f6d7a",
  text_size = 4,
  text_hjust = -0.8,
  text_vjust = -0.5
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{bins}{The number of bins to use in the histogram (default is 150).}

\item{alpha}{The level of transparency for the bars in the histogram
(default is 0.5).}

\item{colors}{A named vector of colors to use for the histograms, 
with "Simulated" and "Observed" as the named elements 
(default is c("Simulated" = "#FB8500", "Observed" = "#36949D")).}

\item{xlim}{Limits for the x axis of the histogram (e.g. c(0.8, 1)).}

\item{vline_color}{Color of the vertical line indicating 
the critical distance with KNN.}

\item{text_size}{Font size for the label of the critical distance.}

\item{text_hjust}{Horizontal justification for the label 
of the critical distance.}

\item{text_vjust}{Vertical justification for the label 
of the critical distance.}
}
\value{
A ggplot object.
}
\description{
This function creates a histogram of the simulated and observed distances 
of a ClusterSet object.
}
\examples{
# Set verbosity to 1 to only display info messages.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative genes
res <- select_genes(m,
                    distance = "kendall",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
plot_dist(res)

}
