% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  object,
  center = TRUE,
  ceil = 1,
  floor = -1,
  cell_clusters = NULL,
  show_dendro = TRUE,
  gene_clusters = "all",
  use_top_genes = TRUE,
  name = NULL,
  xlab = NULL,
  ylab = NULL,
  colorbar_name = "Expression level",
  show_legend = TRUE,
  colors = c("#A9D6E5", "#2166AC", "#000000", "#B2182B", "#FFCA3A"),
  colors_cell_clusters = c("#9F1717", "#AE5B11", "#C48D00", "#517416", "#115C8A",
    "#584178", "#9D1C70", "#E96767", "#EC9342", "#FFCA3A", "#8AC926", "#4DADE8",
    "#9579B9", "#E25CB4", "#DB2020", "#DA7316", "#F0AE00", "#6D9D1E", "#1882C0",
    "#71529A", "#D02494", "#EF9292", "#F2B57D", "#FFDA77", "#B6E36A", "#7BC4EE",
    "#AD98C9", "#EA8AC9"),
  row_labels = TRUE,
  col_labels = FALSE,
  label_size = 9,
  line_size_vertical = 15,
  line_size_horizontal = 15
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{center}{A logical to indicate whether to center row..}

\item{ceil}{A value for ceiling (NULL for no ceiling). Ceiling is performed centering.}

\item{floor}{A value for flooring (NULL for no flooring). Flooring is performed after centering.}

\item{cell_clusters}{A vector of cell clusters with cell barcodes as names.}

\item{show_dendro}{A logical to indicate whether to show column dendrogram.}

\item{gene_clusters}{A cluster id to plot. Default is NULL for plotting all cluster.}

\item{use_top_genes}{A logical to indicate whether to use highly similar genes in the slot top_genes of ClusterSet.}

\item{name}{A title for the heatmap.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{colorbar_name}{A title for the colorbar.}

\item{show_legend}{A logical to indicate whether to show colorbar.}

\item{colors}{A vector of colors.}

\item{colors_cell_clusters}{A vector of colors for column annotations.}

\item{row_labels}{A logical to indicate whether to show row labels.}

\item{col_labels}{A logical to indicate whether to show col labels.}

\item{label_size}{A value for label font size.}

\item{line_size_vertical}{An integer for the size of horizontal white line which separate gene clusters.}

\item{line_size_horizontal}{An integer for the size of vertical white line  which separate cell clusters.}
}
\value{
Iheatmap-class object.
}
\description{
Plot the observed and simulated distance with the Kth nearest neighbors.
}
\examples{
m <- matrix(rnorm(40000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- find_gene_clusters(data=m,
                          distance_method="pearson",
                          av_dot_prod_min = 0,
                          inflation = 2,
                          k = 25,
                          fdr = 10)

plot_heatmap(object = res)
plot_heatmap(object = res, cluster = "1")

}
