% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  object,
  center = TRUE,
  ceil = 2,
  floor = -2,
  cell_order = NULL,
  cell_ordering_method = "hclust",
  name = NULL,
  xlab = NULL,
  ylab = NULL,
  colorbar_name = "Expression level",
  show_legend = TRUE,
  colors = c("#FFCA3A", "#B2182B", "#000000", "#2166AC", "#A9D6E5"),
  row_labels = TRUE,
  col_labels = FALSE,
  line_size = 15
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{center}{A logical to indicate whether to center row..}

\item{ceil}{A value for ceiling (NULL for no ceiling). Ceiling is performed centering.}

\item{floor}{A value for flooring (NULL for no flooring). Flooring is performed after centering.}

\item{cell_order}{A vector of cell names already ordered.}

\item{cell_ordering_method}{The clustering method to be used. This must be "hclust", ADD OTHER CLUSTERING METHOD.}

\item{name}{A title for the heatmap.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{colorbar_name}{A title for the colorbar.}

\item{show_legend}{A logical to indicate whether to show colorbar.}

\item{colors}{A vector of colors.}

\item{row_labels}{A logical to indicate whether to show row labels.}

\item{col_labels}{A logical to indicate whether to show col labels.}

\item{line_size}{An integer for the horizontal white line size.}
}
\value{
Iheatmap-class object
}
\description{
Plot the observed and simulated distance with the Kth nearest neighbors.
}
\examples{
m <- matrix(rnorm(6000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- DBFMCL(data=m,
distance_method="pearson",
av_dot_prod_min = 0,
inflation = 1.2,
k=25,
fdr = 10,
mcl_cmd_line = TRUE)

plot_heatmap(object = res)

}
