% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  object,
  center = TRUE,
  ceil = 1,
  floor = -1,
  cell_clusters = NULL,
  show_dendro = TRUE,
  use_top_genes = FALSE,
  interactive = TRUE,
  name = NULL,
  xlab = NULL,
  ylab = NULL,
  colorbar_name = "Exp. level",
  show_legend = TRUE,
  colors = colors_for_gradient("Ju1"),
  colors_cell_clusters = NULL,
  row_labels = TRUE,
  col_labels = FALSE,
  label_size = 10,
  line_size_vertical = 3,
  line_size_horizontal = 3,
  link = c("average", "complete", "single")
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{center}{A logical to indicate whether to center row..}

\item{ceil}{A value for ceiling (NULL for no ceiling). Ceiling is performed centering.}

\item{floor}{A value for flooring (NULL for no flooring). Flooring is performed after centering.}

\item{cell_clusters}{A vector of cell clusters with cell barcodes as names.}

\item{show_dendro}{A logical to indicate whether to show column dendrogram.}

\item{use_top_genes}{A logical to indicate whether to use highly similar genes in the slot top_genes of ClusterSet.}

\item{interactive}{A logical to indicate if the heatmap should be interactive.}

\item{name}{A title for the heatmap.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{colorbar_name}{A title for the colorbar.}

\item{show_legend}{A logical to indicate whether to show colorbar.}

\item{colors}{A vector of colors.}

\item{colors_cell_clusters}{A named vector of colors for cell identity annotations.}

\item{row_labels}{A logical to indicate whether to show row labels.}

\item{col_labels}{A logical to indicate whether to show col labels.}

\item{label_size}{A value for label font size.}

\item{line_size_vertical}{An integer for the size of horizontal white line which separate gene clusters.}

\item{line_size_horizontal}{An integer for the size of vertical white line  which separate cell clusters.}

\item{link}{The aggloremative criterion for hierarchical clustering. One of "average", "complete" or "single". 
Default to average.}
}
\value{
Iheatmap-class object.
}
\description{
Plot the observed and simulated distance with the Kth nearest neighbors.
}
\examples{
# Set verbosity to 1 to only display info messages.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative genes
res <- select_genes(m,
                    distance = "kendall",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Cluster informative features
res <- gene_clustering(res,
                       inflation = 1.2,
                       keep_nn = FALSE,
                       k = 5)
# Plot heatmap of all gene clusters
plot_heatmap(res)

# Plot heatmap of gene cluster 1
plot_heatmap(res[1,]) 

}
