% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  object,
  center = TRUE,
  ceil = 1,
  floor = -1,
  cell_order = NULL,
  show_dendro = FALSE,
  gene_cluster = NULL,
  use_top_genes = FALSE,
  use_core_cells = FALSE,
  name = NULL,
  xlab = NULL,
  ylab = NULL,
  colorbar_name = "Expression level",
  show_legend = TRUE,
  colors = c("#A9D6E5", "#2166AC", "#000000", "#B2182B", "#FFCA3A"),
  row_labels = TRUE,
  col_labels = FALSE,
  label_size = 9,
  line_size_vertical = 15,
  line_size_horizontal = 15
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{center}{A logical to indicate whether to center row..}

\item{ceil}{A value for ceiling (NULL for no ceiling). Ceiling is performed centering.}

\item{floor}{A value for flooring (NULL for no flooring). Flooring is performed after centering.}

\item{cell_order}{A vector of cell names already ordered.}

\item{show_dendro}{A logical to indicate whether to show column dendrogram.}

\item{gene_cluster}{A cluster id to plot. Default is NULL for plotting all cluster.}

\item{use_top_genes}{A logical to indicate whether to use highly similar genes in the slot top_genes of ClusterSet.}

\item{use_core_cells}{A logical to indicate whether to use core cells obtained by cell_clust function.}

\item{name}{A title for the heatmap.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{colorbar_name}{A title for the colorbar.}

\item{show_legend}{A logical to indicate whether to show colorbar.}

\item{colors}{A vector of colors.}

\item{row_labels}{A logical to indicate whether to show row labels.}

\item{col_labels}{A logical to indicate whether to show col labels.}

\item{label_size}{A value for label font size.}

\item{line_size_vertical}{An integer for the size of horizontal white line which separate gene clusters.}

\item{line_size_horizontal}{An integer for the size of vertical white line  which separate cell clusters.}

\item{cell_ordering_method}{The clustering method to be used. This must be "hclust", ADD OTHER CLUSTERING METHOD.}
}
\value{
Iheatmap-class object
}
\description{
Plot the observed and simulated distance with the Kth nearest neighbors.
}
\examples{
m <- matrix(rnorm(40000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- find_gene_clusters(data=m,
                          distance_method="pearson",
                          av_dot_prod_min = 0,
                          inflation = 2,
                          k = 25,
                          fdr = 10)

plot_heatmap(object = res)
plot_heatmap(object = res, cluster = "1")

}
