% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Plot mean expression profiles of each cluster from a ClusterSet object.}
\usage{
plot_profiles(
  data = NULL,
  ident = NULL,
  nb_column = NULL,
  color_cell_type = NULL,
  size_text_y = 5,
  size_label = 2,
  legend_name = "Cell\\ntype"
)
}
\arguments{
\item{data}{A ClusterSet object.}

\item{ident}{A named vector containing the cell type identities for each cell.
Typically the result from the Idents() function on a Seurat object (see Seurat library).}

\item{nb_column}{The number of columns in the facet grid of the plot. If not provided,
it is automatically computed as the square root of the number of cell types.}

\item{color_cell_type}{A named vector of colors (with cell type as names) used to 
distinguish between different cell types in the plot. If not provided, the default 
hue color palette is used.}

\item{size_text_y}{The font size of the y-axis tick labels.}

\item{size_label}{The font size of the cluster labels.}

\item{legend_name}{A name for the legend.}
}
\value{
A ggplot object showing the expression profiles of cell type-specific genes.
}
\description{
This function generates a barplot showing the expression profiles of
cell type-specific genes across different cell types.
}
\examples{
# Load a Seurat object
load_example_dataset("7870305/files/lymph_node_tiny_2")
# Load a ClusterSet object
load_example_dataset("7870305/files/lymph_node_tiny_clusters_2")
                            
plot_profiles(lymph_node_tiny_clusters_2, ident=Seurat::Idents(lymph_node_tiny_2))
lv <- levels(Seurat::Idents(lymph_node_tiny_2))
pal <- discrete_palette(n=length(lv))
names(pal) <- lv
plot_profiles(lymph_node_tiny_clusters_2, ident=Seurat::Idents(lymph_node_tiny_2), color_cell_type = pal)
plot_profiles(lymph_node_tiny_clusters_2[2:4,], ident=Seurat::Idents(lymph_node_tiny_2), color_cell_type = pal)

}
