% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Plot mean expression profiles of each cluster from a ClusterSet object.}
\usage{
plot_profiles(
  data = NULL,
  ident = NULL,
  nb_column = NULL,
  color_cell_type = NULL,
  size_text_y = 5,
  size_label = 2,
  legend_name = "Cell\\ntype"
)
}
\arguments{
\item{data}{A ClusterSet object.}

\item{ident}{A named vector containing the cell type identities for each cell.
Typically the result from the Idents() function on a Seurat object (see Seurat library).}

\item{nb_column}{The number of columns in the facet grid of the plot. If not provided,
it is automatically computed as the square root of the number of cell types.}

\item{color_cell_type}{A named vector of colors (with cell type as names) used to 
distinguish between different cell types in the plot. If not provided, the default 
hue color palette is used.}

\item{size_text_y}{The font size of the y-axis tick labels.}

\item{size_label}{The font size of the cluster labels.}

\item{legend_name}{A name for the legend.}
}
\value{
A ggplot object showing the expression profiles of cell type-specific genes.
}
\description{
This function generates a barplot showing the expression profiles of
cell type-specific genes across different cell types.
}
\examples{
# Load a Seurat object
data(pbmc_small, package = "SeuratObject")

# Compute the signatures using find_gene_clusters()
clust_set <- select_genes(pbmc_small,
                          distance = "kendall",
                          k = 75,
                          highest = 0.3,
                          fdr = 1e-8,
                          row_sum = -Inf,
                          no_dknn_filter = TRUE)
                          
# Cluster informative features
clust_set <- gene_clustering(clust_set, 
                            inflation = 1.6,
                            keep_nn = FALSE,
                            k = 5)
                            
plot_profiles(clust_set, ident=Seurat::Idents(pbmc_small))
pal <- c("#4E79A7", "#A0CBE8", "#F28E2B")
names(pal) <- levels(Seurat::Idents(pbmc_small))
plot_profiles(clust_set, ident=Seurat::Idents(pbmc_small), color_cell_type = pal)
plot_profiles(clust_set[2:4,], ident=Seurat::Idents(pbmc_small), color_cell_type = pal)

}
