% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial.R
\name{plot_spatial}
\alias{plot_spatial}
\title{XY scatter plot of Visium data with hexagonal spots.}
\usage{
plot_spatial(
  seurat_obj = NULL,
  gene_name = NULL,
  metadata = NULL,
  intensity_slot = c("data", "counts", "sct"),
  title = "",
  size_title = 2,
  face_title = c("plain", "italic", "bold", "bold.italic"),
  legend = TRUE,
  barwidth = 1,
  barheight = 3,
  axis = TRUE,
  pt_size = 3.6,
  pt_shape = 6,
  pt_star = TRUE,
  stroke = 0,
  colours = colors_for_gradient("Ju1"),
  coord_flip = T
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing spatial expression data.}

\item{gene_name}{The name of the gene to plot.}

\item{metadata}{Provide the name of a metadata that will be used instead of genes (i.e. from 
meta.data) slot of a seurat object.}

\item{intensity_slot}{The assay slot to use for the gene expression values.
Must be one of "sct", "counts", or "data". Default is "sct".}

\item{title}{The title of the plot. Default is an empty string.}

\item{size_title}{The size of the title.}

\item{face_title}{Font face for the title. Possible values are “plain”, “italic”, “bold” and “bold.italic”.}

\item{legend}{Whether to display a legend for the color scale. Default is FALSE.}

\item{barwidth}{A numeric or a grid::unit() object specifying the width of the colourbar. 
Default value is legend.key.width or legend.key.size in theme() or theme.}

\item{barheight}{A numeric or a grid::unit() object specifying the height of the colourbar. 
Default value is legend.key.height or legend.key.size in theme() or theme.}

\item{axis}{Whether to display a axis for the color scale. Default is FALSE.}

\item{pt_size}{The size of the points in the plot. Default is 2.1.}

\item{pt_shape}{The shape of the points in the plot. Default is 16 (a circle).}

\item{pt_star}{A boolean. Whether to use ggstar shapes.}

\item{stroke}{The thickness of margin of points.}

\item{colours}{A vector of colors.}

\item{coord_flip}{Whether to flip coordinates.}
}
\value{
A ggplot2 object containing the scatter plot.
}
\description{
Given a Seurat Spatial object, this function creates a scatter plot of the 
spatial expression of a gene across spots, where the X and Y coordinates 
represent the spatial location of spots and the color represents the 
expression level of the gene. This function as been tested with visium 
data at the moment. Defaut shape is hexagon.
}
\examples{
library(SeuratData)
library(Seurat)
if(! "stxBrain" \%in\% InstalledData()$Dataset){ InstallData("stxBrain") }
anterior1 <- LoadData("stxBrain", type = "anterior1")
plot_spatial(seurat_obj = anterior1, gene_name = "Hpca")
anterior1 <- SCTransform(anterior1, assay = "Spatial")
plot_spatial(seurat_obj = anterior1, gene_name = "Hpca", intensity_slot="sct")
plot_spatial(seurat_obj = anterior1, metadata = "nCount_SCT")
}
