% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_msg}
\alias{print_msg}
\title{Print a message based on the level of verbosity}
\usage{
print_msg(msg, msg_type = c("INFO", "DEBUG", "WARNING", "STOP"))
}
\arguments{
\item{msg}{The message to be printed}

\item{msg_type}{The type of message, one of "INFO", "DEBUG", or "WARNING"}
}
\value{
None
}
\description{
Print a message based on the level of verbosity
}
\examples{
opt_warn <- options()$warn
set_verbosity(1)
print_msg("Hello world!", "INFO")
set_verbosity(3)
print_msg("Debugging message", "DEBUG")
set_verbosity(0)
print_msg("Hello world!", "INFO")
print_msg("Debugging message", "DEBUG")
options(warn=0)
print_msg("Warning message", "WARNING")
options(warn=-1)
print_msg("A warning message not displayed", "WARNING")
options(warn=opt_warn)
}
