% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{rename_clust}
\alias{rename_clust}
\title{Rename the gene clusters of a ClusterSet}
\usage{
rename_clust(object, new_names = NULL)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{new_names}{The new names for the clusters.}
}
\description{
Rename the gene clusters of a ClusterSet.
}
\examples{
# load a dataset
data("scigenex_test_I1.2")
new_obj <- rename_clust(scigenex_test_I1.2, new_names = letters[1:nclust(scigenex_test_I1.2)])
}
\keyword{internal}
