% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_genes.R
\name{reorder_genes,ClusterSet-method}
\alias{reorder_genes,ClusterSet-method}
\title{Reorder genes in a ClusterSet object.}
\usage{
\S4method{reorder_genes}{ClusterSet}(
  object = NULL,
  order_by = c("gene_names", "hclust", "correlation"),
  method = c("pearson", "maximum", "manhattan", "canberra", "binary", "euclidean",
    "abspearson", "correlation", "abscorrelation", "spearman", "kendall"),
  link = c("average", "single", "complete", "ward", "mcquitty", "median", "centroid",
    "centroid2"),
  decreasing = FALSE,
  nb_proc = 2
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{order_by}{A character string specifying the ordering rule. 
Default is "gene_names" which reorder genes in alphanumerical 
order. If 'hclust', hierarchical clustering is applied. If 
'correlation' the gene as ordered based on there similarity 
with the cluster center.}

\item{method}{The distance measure to be used order_by=='hclust'.}

\item{link}{The agglomeration method to be used if order_by=='hclust'.}

\item{decreasing}{Whether the sorting should be decreasing (FALSE by default)}

\item{nb_proc}{The number of processor to be used if order_by=='hclust'.}
}
\value{
ClusterSet-class object
}
\description{
Reorder genes in a ClusterSet object according to various
ordering rules.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
                       
# Reorder genes using hierarchical clustering
pbmc3k_medium_clusters <- reorder_genes(pbmc3k_medium_clusters, order_by = "hclust")

# Plot heatmap of gene clusters
plot_heatmap(pbmc3k_medium_clusters)

# Reorder and plot
plot_heatmap(reorder_genes(pbmc3k_medium_clusters[1,], 
                           order_by = "gene_names"), 
             label_size = 5)

}
