% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_genes.R
\name{reorder_genes}
\alias{reorder_genes}
\title{reorder_genes}
\usage{
reorder_genes(object, order_by = c("gene_names", "hclust", "correlation"))
}
\arguments{
\item{object}{A ClusterSet object.}

\item{order_by}{A character string related to the reordering type. 
Default is "correlation" which reorder genes based on their correlation 
within their respective gene cluster.}
}
\value{
ClusterSet-class object
}
\description{
Reorder genes in gene cluster.
}
\examples{
# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative genes
res <- select_genes(m,
                    distance = "kendall",
                    k = 75,
                    highest = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
                    
# Cluster informative features
res <- gene_clustering(res, 
                       inflation = 1.2,
                       keep_nn = FALSE,
                       k = 5)
                       
# Reorder genes using hierarchical clustering
res <- reorder_genes(res, order_by = "hclust")

# Plot heatmap of gene clusters
plot_heatmap(res)

}
