% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_genes.R
\name{reorder_genes}
\alias{reorder_genes}
\title{Reorder genes in a ClusterSet object.}
\usage{
reorder_genes(
  object = NULL,
  order_by = NULL,
  method = NULL,
  link = NULL,
  decreasing = FALSE,
  nb_proc = 2
)
}
\arguments{
\item{object}{A ClusterSet object.}

\item{order_by}{A character string specifying the ordering rule. 
Default is "gene_names" which reorder genes in alphanumerical 
order. If 'hclust', hierarchical clustering is applied. If 
'correlation' the gene as ordered based on there similarity 
with the cluster center.}

\item{method}{The distance measure to be used order_by=='hclust'.}

\item{link}{The agglomeration method to be used if order_by=='hclust'.}

\item{decreasing}{Whether the sorting should be decreasing (FALSE by default)}

\item{nb_proc}{The number of processor to be used if order_by=='hclust'.}
}
\value{
ClusterSet-class object
}
\description{
Reorder genes in a ClusterSet object according to various
ordering rules.
}
\keyword{internal}
