% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_genes.R
\name{reorder_genes}
\alias{reorder_genes}
\title{reorder_genes}
\usage{
reorder_genes(object, order_by = c("gene_names", "hclust", "correlation"))
}
\arguments{
\item{object}{A ClusterSet object.}

\item{order_by}{A character string related to the reordering type. 
Default is "correlation" which reorder genes based on their correlation 
within their respective gene cluster.}
}
\value{
ClusterSet-class object
}
\description{
Reorder genes in gene cluster.
}
\examples{
\dontrun{
m <- matrix(rnorm(80000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2
res <- DBFMCL(data=m,
              distance_method="pearson",
              av_dot_prod_min = 0,
              inflation = 2,
              k=25,
             fdr = 10)

res <- reorder_genes(res, "correlation")

plot_heatmap(res,
             use_core_cells = FALSE,
             use_top_genes = FALSE)

}

}
