% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_genes.R
\name{top_genes}
\alias{top_genes}
\title{Most co-expressed genes from each gene cluster}
\usage{
top_genes(object, top = 20, fast = FALSE, distance_method = NULL)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{top}{A value for the number of genes to select from each cluster.}

\item{fast}{Use qlcMatrix::corSparse for pearson computation. Default to cor(). Default to FALSE for retro-compatibility.}

\item{distance_method}{Overright the object distance_method (slot parameters$distance_method). Useful when object was created using cluster_set_from_matrix() for instance. One of c("kendall", "spearman", "cosine", "euclidean", "pearson") or NULL.}
}
\value{
A \code{ClusterSet} object.
}
\description{
Extract the most highly co-expressed genes of each gene cluster.
}
\keyword{#}
\keyword{1}
\keyword{<-}
\keyword{Load}
\keyword{Set}
\keyword{Store}
\keyword{a}
\keyword{dataset}
\keyword{display}
\keyword{genes}
\keyword{in}
\keyword{info}
\keyword{internal}
\keyword{load_example_dataset('7871581/files/pbmc3k_medium_clusters')}
\keyword{messages}
\keyword{object}
\keyword{only.}
\keyword{pbmc3k_medium_clusters}
\keyword{set_verbosity(1)}
\keyword{the}
\keyword{to}
\keyword{top}
\keyword{top_genes(pbmc3k_medium_clusters)}
\keyword{verbosity}
