% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_genes.R
\name{top_genes}
\alias{top_genes}
\title{Best co-expressed genes from each gene cluster}
\usage{
top_genes(object, top = 20, cluster = "all")
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{top}{A value for the number of the most similar genes in gene clusters.}

\item{cluster}{A vector of gene cluster identity.}
}
\value{
A \code{ClusterSet} object.
}
\description{
Extract the highly co-expressed genes of each gene cluster.
}
\examples{

m <- create_3_rnd_clust()

res <- find_gene_clusters(data=m,
                             distance_method="pearson",
                             inflation = 2,
                             k=75,
                             row_sum=-Inf,
                             highest=0.3,
                             min_nb_supporting_cell = 0,
                             fdr = 1e-8)
              
res <- top_genes(object = res, top=500, cluster="all")
res@top_genes

\dontrun{
## with an artificial dataset

m <- matrix(rnorm(80000), nc = 20)
res <- get_data_for_scigenex(data=m)
}

}
