% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{top_genes}
\alias{top_genes}
\title{Best co-expressed genes from each gene cluster}
\usage{
top_genes(object, top = 20, cluster = "all")
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{top}{A value for the number of the most similar genes in gene clusters.}

\item{cluster}{A vector of gene cluster identity.}
}
\value{

}
\description{
Extract the highly co-expressed genes of each gene cluster.
}
\examples{

m <- matrix(rnorm(80000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- DBFMCL(data=m,
              distance_method="pearson",
              av_dot_prod_min = 0,
              inflation = 1.2,
              k=25,
              fdr = 10)
              
res <- top_genes(object = res, top=500, cluster="all")
res@top_genes

\dontrun{
## with an artificial dataset

m <- matrix(rnorm(80000), nc = 20)
res <- get_data_4_DBFMCL(data=m)
}

}
