% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_genes.R
\name{top_genes}
\alias{top_genes}
\title{Most co-expressed genes from each gene cluster}
\usage{
top_genes(object, top = 20)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{top}{A value for the number of genes to select from each cluster.}
}
\value{
A \code{ClusterSet} object.
}
\description{
Extract the most highly co-expressed genes of each gene cluster.
}
\keyword{internal}
