% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{viz_enrich,ClusterSet-method}
\alias{viz_enrich,ClusterSet-method}
\title{Visualization of enrichment analyses results from a ClusterSet object}
\usage{
\S4method{viz_enrich}{ClusterSet}(
  object,
  clusters = "all",
  type = c("dotplot", "barplot"),
  nb_terms = 20,
  terms_size = 50,
  font_size = 4
)
}
\arguments{
\item{object}{A \code{ClusterSet} object.}

\item{clusters}{A vector of cluster id to plot.}

\item{type}{A vector providing the type of plot.}

\item{nb_terms}{An integer indicating the number of terms in the plot.}

\item{terms_size}{An interger indicating the wrap length of terms}

\item{font_size}{The font size for the x axis.}
}
\value{
A \code{ClusterSet} object
}
\description{
Display enrichment analyses results for a ClusterSet object.
}
\examples{
library(Seurat)
# Load a Seurat object
data(pbmc_small, package = "SeuratObject")

# Compute the signatures using find_gene_clusters()
clust_set <- select_genes(pbmc_small,
                          distance = "pearson",
                          no_dknn_filter = TRUE)
                          
# Cluster informative features
clust_set <- gene_clustering(clust_set, 
                            inflation = 2)
# Plot 
clust_set <- top_genes(clust_set)
plot_ggheatmap(clust_set[, names(Idents(pbmc_small))], ident = Seurat::Idents(pbmc_small))

# Do enrichment analysis using GO ontology
# Only for cluster 1
clust_set <- enrich_go(clust_set[1:2,])

viz_enrich(clust_set, cluster = 1, nb_terms = 5)
}
