% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterSet_class.R
\name{which_clust}
\alias{which_clust}
\title{Which clusters contain a set of genes.}
\usage{
which_clust(object, genes = NULL)
}
\arguments{
\item{object}{a ClusterSet object.}

\item{genes}{The genes to be searched in the ClusterSet.}
}
\description{
Which clusters contain a set of genes.
}
\examples{
# load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
hit <- which_clust(pbmc3k_medium_clusters, genes = c("TJP2", "GLA", "UNKNOWN"))
}
\keyword{internal}
