% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{write_clust}
\alias{write_clust}
\alias{write_clust,ClusterSet-method}
\title{write_clust}
\usage{
write_clust(object, filename_out = NULL, out_path = ".", nb_na_row = 3)

\S4method{write_clust}{ClusterSet}(object, filename_out = NULL, out_path = ".", nb_na_row = 5)
}
\arguments{
\item{object}{ClusterSet.}

\item{filename_out}{The outfile name.}

\item{out_path}{The path to the file.}

\item{nb_na_row}{Number of separating rows (containing NAs).}
}
\value{
Write a file.
}
\description{
Write a ClusterSet into a flat file.
}
\examples{
# see online help.
}
