% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfmcl_main.R
\name{write_dbf}
\alias{write_dbf}
\alias{write_dbf,DBFMCLresult-method}
\title{Title
Write a DBFMCLresult into a flat file.}
\usage{
write_dbf(object, filename_out = NULL, path = ".", verbose = TRUE)

\S4method{write_dbf}{DBFMCLresult}(object, filename_out = NULL, path = ".", verbose = TRUE)
}
\arguments{
\item{object}{DBFMCLresult.}

\item{filename_out}{The outfile name.}

\item{path}{The path to the file.}

\item{verbose}{Whether verbosity should be activated.}
}
\value{

}
\description{
Title
Write a DBFMCLresult into a flat file.
}
